### 2022-05-26 #4427 @editorialbot
[PRE REVIEW]: GPJax: A didactic Gaussian process package in Jax
**Submitting author:** @thomaspinder (<a href="http://orcid.org/0000-0001-9854-4265">Thomas Pinder</a>)
**Repository:** https://github.com/thomaspinder/GPJax
**Branch with paper.md** (empty if default branch):
**Version:** v0.4.5
