### 2022-05-26 26f481c
Update sparse_gps.py
Fix bug in SVGP for covariates higher than a single dimension.
