### 2022-05-31 790ea25
Fix memory bug in SVGP.
Fix memory bug.

Broadcasting was wrong way round.
