+++ requirements.txt
@@ -7,3 +7,4 @@ tensorflow == 2.8.1
 tensorflow-probability==0.16.0
 tqdm>=4.0.0
 ml-collections==0.1.0
+protobuf==3.19.0
