+++ docs/nbs/spectral_kernels.pct.py
@@ -1,0 +1,58 @@
+# ---
+# jupyter:
+#   jupytext:
+#     cell_metadata_filter: -all
+#     custom_cell_magics: kql
+#     text_representation:
+#       extension: .py
+#       format_name: percent
+#       format_version: '1.3'
+#       jupytext_version: 1.11.2
+#   kernelspec:
+#     display_name: Python 3.9.7 ('gpjax')
+#     language: python
+#     name: python3
+# ---
+
+# %%
+import gpjax as gpx
+import jax
+import jax.numpy as jnp
+import jax.random as jr
+import matplotlib.pyplot as plt
+import optax as ox
+import distrax as dx
+
+key = jr.PRNGKey(123)
+
+# %%
+n = 100
+noise = 0.3
+
+x = jr.uniform(key=key, minval=-3.0, maxval=3.0, shape=(n,)).sort().reshape(-1, 1)
+f = lambda x: jnp.sin(4 * x) + jnp.cos(2 * x)
+signal = f(x)
+y = signal + jr.normal(key, shape=signal.shape) * noise
+
+D = gpx.Dataset(X=x, y=y)
+
+xtest = jnp.linspace(-3.25, 3.25, 500).reshape(-1, 1)
+ytest = f(xtest)
+
+# %%
+fig, ax = plt.subplots(figsize=(10, 5))
+ax.plot(xtest, ytest, label="Latent function")
+ax.plot(x, y, "o", label="Observations")
+ax.legend(loc="best")
+
+# %%
+base_kernel = gpx.Matern32()
+kernel = gpx.RandomFourierFeature(key = key, base_kernel = base_kernel, num_basis_fns=50)
+prior = gpx.Prior(kernel=kernel)
+
+# %%
+kernel.params
+
+# %%
+
+# %%
