+++ docs/nbs/classification.pct.py
@@ -44,7 +44,7 @@ y = 0.5 * jnp.sign(jnp.cos(3 * x + jr.normal(key, shape=x.shape) * 0.05)) + 0.5
 
 D = gpx.Dataset(X=x, y=y)
 
-xtest = jnp.linspace(-1., 1., 500).reshape(-1, 1)
+xtest = jnp.linspace(-1.0, 1.0, 500).reshape(-1, 1)
 plt.plot(x, y, "o", markersize=8)
 # %% [markdown]
 # ## MAP inference
@@ -60,7 +60,7 @@ likelihood = gpx.Bernoulli(num_datapoints=D.n)
 posterior = prior * likelihood
 print(type(posterior))
 # %% [markdown]
-# Whilst the latent function is Gaussian, the posterior distribution is non-Gaussian since our generative model first samples the latent GP and propagates these samples through the likelihood function's inverse link function. This step prevents us from being able to analytically integrate the latent function's values out of our posterior, and we must instead adopt alternative inference techniques. We begin with maximum a posteriori (MAP) estimation, a fast inference procedure to obtain point estimates for the latent function and the kernel's hyperparameters by maximising the marginal log-likelihood. 
+# Whilst the latent function is Gaussian, the posterior distribution is non-Gaussian since our generative model first samples the latent GP and propagates these samples through the likelihood function's inverse link function. This step prevents us from being able to analytically integrate the latent function's values out of our posterior, and we must instead adopt alternative inference techniques. We begin with maximum a posteriori (MAP) estimation, a fast inference procedure to obtain point estimates for the latent function and the kernel's hyperparameters by maximising the marginal log-likelihood.
 # %% [markdown]
 # To begin we obtain a set of initial parameter values through the `initialise` callable, and transform these to the unconstrained space via `transform` (see the [regression notebook](https://gpjax.readthedocs.io/en/latest/nbs/regression.html)). We also define the negative marginal log-likelihood, and JIT compile this to accelerate training.
 # %%
@@ -96,16 +96,15 @@ map_estimate = gpx.fit(
 # %%
 # Adapted from BlackJax's introduction notebook.
 num_adapt = 1000
-num_samples = 1000
+num_samples = 500
 
 mll = jax.jit(posterior.marginal_log_likelihood(D, constrainer, negative=False))
 
-adapt = blackjax.window_adaptation(
-    blackjax.nuts, mll, num_adapt, target_acceptance_rate=0.65
-)
+adapt = blackjax.window_adaptation(blackjax.nuts, mll, num_adapt, target_acceptance_rate=0.65)
 
 # Initialise the chain
 last_state, kernel, _ = adapt.run(key, params)
+
 
 def inference_loop(rng_key, kernel, initial_state, num_samples):
     def one_step(state, rng_key):
@@ -133,9 +132,9 @@ print(f"Acceptance rate: {acceptance_rate:.2f}")
 # Our acceptance rate is slightly too large, prompting an examination of the chain's trace plots. A well-mixing chain will have very few (if any) flat spots in its trace plot whilst also not having too many steps in the same direction. In addition to the model's hyperparameters, there will be 500 samples for each of the 100 latent function values in the `states.position` dictionary. We depict the chains that correspond to the model hyperparameters and the first value of the latent function for brevity.
 # %%
 fig, (ax0, ax1, ax2) = plt.subplots(ncols=3, figsize=(15, 5), tight_layout=True)
-ax0.plot(states.position['kernel']['lengthscale'])
-ax1.plot(states.position['kernel']['variance'])
[…]
