### 2022-06-01 a5e50ad
Renaming Variational Inference
