+++ gpjax/__init__.py
@@ -20,7 +20,7 @@ from .kernels import (
 from .likelihoods import Bernoulli, Gaussian
 from .mean_functions import Constant, Zero
 from .parameters import copy_dict_structure, initialise, transform
-from .sparse_gps import SVGP
+from .sparse_gps import StochasticVI
 from .types import Dataset
 from .variational import VariationalGaussian, WhitenedVariationalGaussian
 
