+++ gpjax/variational_families.py
@@ -83,7 +83,6 @@ class VariationalGaussian(AbstractVariationalFamily):
                 }
         )
 
-    # Compute KL divergence at inducing points, KL[q(u)||p(u)]:
     def prior_kl(self, params: Dict) -> Array:
         """Compute the KL-divergence between our current variational approximation and the Gaussian process prior.
 
@@ -149,7 +148,8 @@ class VariationalGaussian(AbstractVariationalFamily):
 class WhitenedVariationalGaussian(VariationalGaussian):
     """The variational Gaussian family of probability distributions."""
 
-    # Compute KL divergence at inducing points, KL[q(u)||p(u)]:
+    name: str = "Whitened Gaussian"
+
     def prior_kl(self, params: Dict) -> Array:
         """Compute the KL-divergence between our current variational approximation and the Gaussian process prior.
 
@@ -196,6 +196,121 @@ class WhitenedVariationalGaussian(VariationalGaussian):
             V = jnp.matmul(A.T, sqrt)
             
             mean = μt + jnp.matmul(A.T, mu)
+            covariance = Ktt - jnp.matmul(A.T, A) + jnp.matmul(V, V.T)
+
+            return dx.MultivariateNormalFullCovariance(
+                jnp.atleast_1d(mean.squeeze()), covariance
+            )
+
+        return predict_fn
+    
+
+@dataclass
+class NaturalVariationalGaussian(AbstractVariationalFamily):
+    """The variational Gaussian family of probability distributions."""
+    prior: Prior
+    inducing_inputs: Array
+    name: str = "Natural Gaussian"
+    natural_vector: Optional[Array] = None
+    natural_matrix: Optional[Array] = None
+    jitter: Optional[float] = DEFAULT_JITTER
+
+    def __post_init__(self):
+        """Initialise the variational Gaussian distribution."""
+        self.num_inducing = self.inducing_inputs.shape[0]
+        add_parameter("inducing_inputs", Identity)
+
+        m = self.num_inducing
+
+        if self.natural_vector is None:
+            self.natural_vector = jnp.zeros((m, 1))
+            add_parameter("natural_vector", Identity)
+
+        if self.natural_matrix is None:
+            self.natural_matrix = -.5 * I(m)
+            add_parameter("natural_matrix", Identity)
+
+    @property
+    def params(self) -> Dict:
+        """Return the natural vector and matrix, inducing inputs, and hyperparameters that parameterise the natural Gaussian distribution."""
+        return concat_dictionaries(
+            self.prior.params, {
+            "variational_family": {
+                "inducing_inputs": self.inducing_inputs,
+                "natural_vector": self.natural_vector,
+                "natural_matrix": self.natural_matrix}
+                }
+        )
+
+    def prior_kl(self, params: Dict) -> Array:
+        """Compute the KL-divergence between our current variational approximation and the Gaussian process prior.
+
+        Args:
+            params (Dict): The parameters at which our variational distribution and GP prior are to be evaluated.
+
+        Returns:
+            Array: The KL-divergence between our variational approximation and the GP prior.
+        """
+        natural_vector = params["variational_family"]["natural_vector"]
+        natural_covariance = params["variational_family"]["natural_covariance"]
+        z = params["variational_family"]["inducing_inputs"]
+        m = self.num_inducing
+        
+        S_inv = -2 * natural_covariance
+        S_inv += I(m) * self.jitter
+        L_inv = jnp.linalg.cholesky(S_inv)
+        L = jsp.linalg.solve_triangular(L_inv, I(m), lower=True)
+        
[…]
