### 2022-06-01 dda0c6d
Natural gradients.
Added Natural Variational Gaussian Family.

NEED TO WRITE TESTS.

The variational tests for variational families are a mess --- much of the testing is done in sparse_gps.py. In order to write these, it might be worth fixing these too.
