+++ gpjax/__init__.py
@@ -21,7 +21,6 @@ from .likelihoods import Bernoulli, Gaussian
 from .mean_functions import Constant, Zero
 from .parameters import copy_dict_structure, initialise, transform
 from .types import Dataset
-
 from .variational_inference import StochasticVI
 from .variational_families import VariationalGaussian, WhitenedVariationalGaussian, NaturalVariationalGaussian
 
