### 2022-06-01 8609155
Rename files.
Renaming sparse_gps.py to variational_inference.py and variational.py to variational_families for structure improvements.

To do: Need to restructure the tests.
