### 2022-06-01 97a1da4
Fix bugs and update tests.
Bug with single dimension in Distrax  multivariate normal is fixed for variational families.

The tests have been updated.
