+++ tests/test_gp.py
@@ -4,7 +4,6 @@ import distrax as dx
 import jax.numpy as jnp
 import jax.random as jr
 import pytest
-import tensorflow_probability.substrates.jax as tfp
 
 from gpjax import Dataset, initialise, likelihoods, transform
 from gpjax.gps import (
