+++ tests/test_mean_functions.py
@@ -2,7 +2,6 @@ import typing as tp
 
 import jax.numpy as jnp
 import pytest
-from numpy import isin
 
 from gpjax.mean_functions import Constant, Zero
 from gpjax.parameters import initialise
