### 2022-06-01 9a7d5be
Update test_gp.py
