### 2022-06-01 dbaf04b
Update tests + fix minor bugs.
