+++ tests/test_variational_families.py
@@ -165,14 +165,38 @@ def test_collapsed_variational_gaussian(n_test, n_inducing, n_datapoints, point_
     assert mu.shape == (n_test,)
     assert sigma.shape == (n_test, n_test)
 @pytest.mark.parametrize("n_inducing", [1, 10, 20])
-def test_natural_variational_gaussian_params(n_inducing):
+def test_natural_variational_gaussian(n_inducing, n_test):
     prior = gpx.Prior(kernel=gpx.RBF())
-    inducing_points = jnp.linspace(-3.0, 3.0, n_inducing).reshape(-1, 1)
+    
+    inducing_inputs = jnp.linspace(-5.0, 5.0, n_inducing).reshape(-1, 1)
+    test_inputs = jnp.linspace(-5.0, 5.0, n_test).reshape(-1, 1)
+
+
     variational_family = gpx.variational_families.NaturalVariationalGaussian(
         prior=prior,
-        inducing_inputs=inducing_points
+        inducing_inputs=inducing_inputs
     )
 
+    # Test init
+    assert variational_family.num_inducing == n_inducing
+
+    assert jnp.sum(variational_family.natural_vector) == 0.0
+    assert variational_family.natural_vector.shape == (n_inducing, 1)
+
+    assert variational_family.natural_matrix.shape == (
+        n_inducing,
+        n_inducing,
+    )
+    assert jnp.all(jnp.diag(variational_family.natural_matrix) == -.5)
+
+    params = gpx.config.get_defaults()
+    assert "variational_root_covariance" in params["transformations"].keys()
+    assert "variational_mean" in params["transformations"].keys()
+
+    assert (variational_family.natural_matrix == -.5 * jnp.eye(n_inducing)).all()
+    assert (variational_family.natural_vector == jnp.zeros((n_inducing, 1))).all()
+
+    # params
     params = variational_family.params
     assert isinstance(params, dict)
     assert "inducing_inputs" in params["variational_family"].keys()
@@ -193,3 +217,24 @@ def test_natural_variational_gaussian_params(n_inducing):
 
     assert (variational_family.natural_matrix == -.5 * jnp.eye(n_inducing)).all()
     assert (variational_family.natural_vector == jnp.zeros((n_inducing, 1))).all()
+
+
+    #Test KL
+    params = variational_family.params
+    kl = variational_family.prior_kl(params)
+    assert isinstance(kl, jnp.ndarray)
+
+    # Test predictions
+    predictive_dist_fn = variational_family(params)
+    assert isinstance(predictive_dist_fn, tp.Callable)
+
+    predictive_dist = predictive_dist_fn(test_inputs)
+    assert isinstance(predictive_dist, dx.Distribution)
+
+    mu = predictive_dist.mean()
+    sigma = predictive_dist.covariance()
+
+    assert isinstance(mu, jnp.ndarray)
+    assert isinstance(sigma, jnp.ndarray)
+    assert mu.shape == (n_test,)
+    assert sigma.shape == (n_test, n_test)
