### 2022-06-01 048b0cc
Update variational_families.py
