### 2022-06-01 35d7036
Update variational_families.py
