### 2022-06-01 40f9dd2
Update variational_families.py
