### 2022-06-01 a0b5ab2
Update variational_families.py
