### 2022-06-01 4399571
Update variational_families.py
