### 2022-06-04 #4455 @editorialbot
[REVIEW]: GPJax
**Submitting author:** @thomaspinder (<a href="http://orcid.org/0000-0001-9854-4265">Thomas Pinder</a>)
**Repository:** https://github.com/thomaspinder/GPJax
**Branch with paper.md** (empty if default branch):
**Version:** 0.4.9
