### 2022-06-08 3874b90
Update variational_families.py
