+++ gpjax/variational_families.py
@@ -44,7 +44,12 @@ class AbstractVariationalFamily:
 
 @dataclass
 class VariationalGaussian(AbstractVariationalFamily):
-    """The variational Gaussian family of probability distributions."""
+    """The variational Gaussian family of probability distributions.
+    
+    The variational family is q(f(.)) = ∫ p(f(.)|u) q(u) du, where u = f(z) are the function values at the inducing inputs z
+    and the distribution over the inducing inputs is q(u) = N(μ, S). We parameterise this over μ and sqrt with S = sqrt sqrt^T.
+
+    """
     prior: Prior
     inducing_inputs: Array
     name: str = "Gaussian"
@@ -83,9 +88,11 @@ class VariationalGaussian(AbstractVariationalFamily):
                 }
         )
 
-    # Compute KL divergence at inducing points, KL[q(u)||p(u)]:
     def prior_kl(self, params: Dict) -> Array:
-        """Compute the KL-divergence between our current variational approximation and the Gaussian process prior.
+        """Compute the KL-divergence between our variational approximation and the Gaussian process prior.
+
+        For this variational family, we have KL[q(f(.))||p(.)] = KL[q(u)||p(u)] = KL[ N(μ, S) || N(μz, Kzz) ],
+        where u = f(z) and z are the inducing inputs. 
 
         Args:
             params (Dict): The parameters at which our variational distribution and GP prior are to be evaluated.
@@ -108,7 +115,11 @@ class VariationalGaussian(AbstractVariationalFamily):
         return qu.kl_divergence(pu)
 
     def predict(self, params: dict) -> Callable[[Array], dx.Distribution]:
-        """Compute the predictive distribution of the GP at the test inputs.
+        """Compute the predictive distribution of the GP at the test inputs t.
+
+        This is the integral q(f(t)) = ∫ p(f(t)|u) q(u) du, which can be computed in closed form as
+          
+            N[f(.); μt + Ktz Kzz^{-1} (μ - μz),  Ktt - Ktz Kzz^{-1} Kzt + Ktz Kzz^{-1} S Kzz^{-1} Kzt ].
 
         Args:
             params (dict): The set of parameters that are to be used to parameterise our variational approximation and GP.
@@ -131,12 +142,21 @@ class VariationalGaussian(AbstractVariationalFamily):
             Ktt = gram(self.prior.kernel, t, params["kernel"])
             Kzt = cross_covariance(self.prior.kernel, z, t, params["kernel"])
             μt = self.prior.mean_function(t, params["mean_function"])
-            A = jsp.linalg.solve_triangular(Lz, Kzt, lower=True)
-            B = jsp.linalg.solve_triangular(Lz.T, A, lower=False)
-            V = jnp.matmul(B.T, sqrt)
+
+            # Lz^{-1} Kzt
+            Lz_inv_Kzt = jsp.linalg.solve_triangular(Lz, Kzt, lower=True)
+
+            # Kzz^{-1} Kzt
+            Kzz_inv_Kzt = jsp.linalg.solve_triangular(Lz.T, Lz_inv_Kzt, lower=False)
+
+            # Ktz Kzz^{-1} sqrt
+            Ktz_Kzz_inv_sqrt = jnp.matmul(Kzz_inv_Kzt.T, sqrt)
             
-            mean = μt + jnp.matmul(B.T, mu - μz)
-            covariance = Ktt - jnp.matmul(A.T, A) + jnp.matmul(V, V.T)
+            # μt + Ktz Kzz^{-1} (μ - μz)
[…]
