### 2022-06-08 0f1f8ef
Update variational_families.py
