### 2022-06-08 340529c
Update gps.py
Improve maths of non-conjugate posterior.
