+++ gpjax/gps.py
@@ -244,11 +244,12 @@ class NonConjugatePosterior(AbstractPosterior):
             Ktx = cross_covariance(self.prior.kernel, t, x, params["kernel"])
             Ktt = gram(self.prior.kernel, t, params["kernel"]) + I(nt) * self.jitter
             μt = self.prior.mean_function(t, params["mean_function"])
-            A = solve_triangular(Lx, Ktx.T, lower=True)
-            latent_var = jnp.diag(jnp.diag(Ktt - jnp.sum(jnp.square(A), -2)))
-            latent_mean = μt + jnp.matmul(A.T, params["latent"])
+            Lx_inv_Kxt = solve_triangular(Lx, Ktx.T, lower=True)
+            latent_var = jnp.diag(Ktt - jnp.matmul(Lx_inv_Kxt.T, Lx_inv_Kxt))
+            latent_mean = μt + jnp.matmul(Lx_inv_Kxt.T, params["latent"])
+
             return dx.MultivariateNormalFullCovariance(
-                jnp.atleast_1d(latent_mean.squeeze()), latent_var
+                jnp.atleast_1d(latent_mean.squeeze()), jnp.diag(latent_var)
             )
 
         return predict_fn
