### 2022-06-08 ee4bd1a
Update variational_families.py
Add maths description and add expectation parameteriation.
