+++ gpjax/variational_families.py
@@ -207,7 +207,7 @@ class WhitenedVariationalGaussian(VariationalGaussian):
 
 @dataclass
 class NaturalVariationalGaussian(AbstractVariationalFamily):
-    """The variational Gaussian family of probability distributions."""
+    """The natural variational Gaussian family of probability distributions."""
     prior: Prior
     inducing_inputs: Array
     name: str = "Natural Gaussian"
@@ -291,12 +291,20 @@ class NaturalVariationalGaussian(AbstractVariationalFamily):
         z = params["variational_family"]["inducing_inputs"]
         m = self.num_inducing
         
+        # S⁻¹ = -2θ₂
         S_inv = -2 * natural_matrix
         S_inv += I(m) * self.jitter
-        L_inv = jnp.linalg.cholesky(S_inv)
-        C = jsp.linalg.solve_triangular(L_inv, I(m), lower=True)
+
+        # S⁻¹ = LLᵀ
+        L = jnp.linalg.cholesky(S_inv)
+
+        # C = L⁻¹I
+        C = jsp.linalg.solve_triangular(L, I(m), lower=True)
         
+        # S = CᵀC
         S = jnp.matmul(C.T, C)
+
+        # μ = Sθ₁
         mu = jnp.matmul(S, natural_vector)
 
         Kzz = gram(self.prior.kernel, z, params["kernel"])
@@ -309,12 +317,144 @@ class NaturalVariationalGaussian(AbstractVariationalFamily):
             Ktt = gram(self.prior.kernel, t, params["kernel"])
             Kzt = cross_covariance(self.prior.kernel, z, t, params["kernel"])
             μt = self.prior.mean_function(t, params["mean_function"])
-            A = jsp.linalg.solve_triangular(Lz, Kzt, lower=True)
-            B = jsp.linalg.solve_triangular(Lz.T, A, lower=False)
-            V = jnp.matmul(B.T, C.T)
+
+            # Lz⁻¹ Kzt
+            Lz_inv_Kzt = jsp.linalg.solve_triangular(Lz, Kzt, lower=True)
+
+            # Kzz⁻¹ Kzt
+            Kzz_inv_Kzt = jsp.linalg.solve_triangular(Lz.T, Lz_inv_Kzt, lower=False)
+
+            # Ktz Kzz⁻¹ Cᵀ
+            Ktz_Kzz_inv_CT = jnp.matmul(Kzz_inv_Kzt.T, C.T)
             
-            mean = μt + jnp.matmul(B.T, mu - μz)
-            covariance = Ktt - jnp.matmul(A.T, A) + jnp.matmul(V, V.T)
+            # μt  +  Ktz Kzz⁻¹ (μ  -  μz)
+            mean = μt + jnp.matmul(Kzz_inv_Kzt.T, mu - μz)
+
+            # Ktt  -  Ktz Kzz⁻¹ Kzt  +  Ktz Kzz⁻¹ S Kzz⁻¹ Kzt  [recall S = CᵀC]
+            covariance = Ktt - jnp.matmul(Lz_inv_Kzt.T, Lz_inv_Kzt) + jnp.matmul(Ktz_Kzz_inv_CT, Ktz_Kzz_inv_CT.T)
+
+            return dx.MultivariateNormalFullCovariance(
+                jnp.atleast_1d(mean.squeeze()), covariance
+            )
+
+        return predict_fn
+
+
+@dataclass
+class ExpectationVariationalGaussian(AbstractVariationalFamily):
+    """The variational Gaussian family of probability distributions."""
+    prior: Prior
+    inducing_inputs: Array
+    name: str = "Natural Gaussian"
+    expectation_vector: Optional[Array] = None
+    expectation_matrix: Optional[Array] = None
[…]
