### 2022-06-08 426e6cb
Update variational_families.py
Add maths description and add expectation parameteriation.
