### 2022-06-08 0604cdd
Update variational_families.py
