### 2022-06-08 5e33437
Update variational_families.py
