### 2022-06-08 700acfb
Update test_variational_families.py
Add test for expectation parameterisation.
