### 2022-06-08 e295182
Update test_variational_families.py
Add test for expectation parameterisation.
