### 2022-06-08 a96281f
Natural gradients sketch.
A rough initial schematic of implementing natural gradients.
