+++ gpjax/parameters.py
@@ -172,6 +172,19 @@ def unconstrain(params: tp.Dict, bijectors: tp.Dict) -> tp.Dict:
 
     return jax.tree_util.tree_map(map, params, bijectors)
 
+def build_identity(params: tp.Dict) -> tp.Dict:
+    """"
+    Args:
+        params (tp.Dict): The parameter set for which trainable statuses should be derived from.
+
+    Returns:
+        tp.Dict: A dictionary of identity forward/backward bijectors. The dictionary is equal in structure to the input params dictionary.
+    """
+    # Copy dictionary structure
+    prior_container = deepcopy(params)
+
+    return jax.tree_map(lambda _: Identity.forward, prior_container)
+
 
 ################################
 # Priors
