### 2022-06-08 4d9f86f
Improve how the maths is written.
