### 2022-06-09 7fabb2a
Improve efficiency of GPR.
Improve efficiency of GPR when predicting both mean and covariance.
