### 2022-06-09 b040760
Consistent jitter to all predict methods
