### 2022-06-09 de73c32
Gotchas notebook
