+++ docs/nbs/gotchas.pct.py
@@ -1,0 +1,75 @@
+# ---
+# jupyter:
+#   jupytext:
+#     cell_metadata_filter: -all
+#     custom_cell_magics: kql
+#     text_representation:
+#       extension: .py
+#       format_name: percent
+#       format_version: '1.3'
+#       jupytext_version: 1.11.2
+#   kernelspec:
+#     display_name: Python 3.9.7 ('gpjax')
+#     language: python
+#     name: python3
+# ---
+
+# %% [markdown]
+# # Common Gotchas in GPJax
+#
+# GPJax is an incredibly flexible framework due to it's low-level of abstraction. However, by providing a large degree of flexibility to users, it is possbile for pain points to be encountered when developing custom models. In this notebook we try to shed some light on these _gothcas_, providing explanations for their presence and giving appropriate solutions where applicable.
+#
+# We will not attempt to cover all Jax based edge cases in this notebook as they are covered superbly in the [Jax - The Sharp Bits](https://jax.readthedocs.io/en/latest/notebooks/Common_Gotchas_in_JAX.html) notebook.
+#
+# _This list is non-exhaustive. If you encounter your own confusion, then please consider opening [contributing](https://github.com/thomaspinder/GPJax/blob/master/CONTRIBUTING.md) a PR to this notebook._
+
+# %% [markdown]
+# ## Chex Dataclasses
+#
+# Dataclasses are supported in base Python since version 3.7. However, for a dataclass object to work with Jax, they must be a registered PyTree. This is where [Chex](https://github.com/deepmind/chex).
+
+# %%
+import gpjax as gpx
+import jax 
+from chex import dataclass
+
+# %% [markdown]
+# ## Parameter transforms
+#
+# Parameters such as the kernel lengthscale or variance have supports on a subset of the real line. Consequently, as we approach the set's boundary during gradient-based optimisation, it becomes possible that we could step outside of the set's support and consequently introduce a numerical and mathematical error into our model. As an example of this, consider the variance paramater which we know must be strictly positive. At the $t^{\text{th}}$ if we had specify a learning rate 
+#
+# A simple, yet highly impractical, solution to this would be to use tiny learning rates, that way we'd reduce our chance of stepping outside of the parameter's support. However, this does not completely remove the problem and it will also be incredibly costly. An alternative solution is to apply a functional mapping to the parameter that projects it from a constrained subspace of the real line onto the full real line. Gradient updates are then performed on the unconstrained parameter value before its value is projected back onto its original support. Such a transformation is known as a bijection. 
+#
+# In GPJax we supply bijective functions using [Distrax](https://github.com/deepmind/distrax). However, __the user is required to transform parameters within the optimisation routine__ through the following syntax
+# ```python
+# unconstrained_params = gpx.transform(constrained_params, unconstrainers)
+# ```
+# The constrained parameters and associated bijections are returned when a user initialises their model using `gpx.initialise`.
+
+# %% [markdown]
+# ## Cholesky factors
+#
+# > "_Symmetric positive definiteness is one of the highest accolades to which a matrix can aspire_" - Nicholas Highman, Accuracy and stability of numerical algorithms
+#
+#
+# ### Why should we care about Cholesky factors?
+#
+# The covariance matrix of a kernel is a symmetric positive definite matrix. As such, we have a range of tools at our disposal to make subsequent operations on the covariance matrix faster. One of these tools is the Cholesky factorisation that uniquely decomposes any symmetric positive-definite matrix $\Sigma$ by
+# $$\Sigma = \mathbf{L}\mathbf{L}^{\top}$$
+# where $\mathbf{L}$ is a lower-triangular matrix.
+#
[…]
