+++ docs/latex_symbols.tex
@@ -1,3 +1,16 @@
 \newcommand{\bbR}{\mathbb{R}}
 \DeclareMathOperator*{\argmax}{arg\,max}
 \DeclareMathOperator*{\argmin}{arg\,min}
+
+\newcommand{\x}{\mathbf{x}}
+\newcommand{\X}{\mathbf{\MakeUppercase{X}}}
+\newcommand{\xx}{\mathbf{x}^{\star}}
+\newcommand{\Xx}{\mathbf{X}^{\star}}
+\newcommand{\Y}{\mathbf{\MakeUppercase{Y}}}
+\newcommand{\y}{\mathbf{y}}
+\newcommand{\by}{\mathbf{y}}
+\newcommand{\chol}{\operatorname{chol}}
+\newcommand{\GP}{\mathcal{GP}}
+
+\newcommand{\Kff}{\mathbf{\MakeUppercase{K}}_{\mathbf{ff}}}
+\newcommand{\IdentityMatrix}{\mathbf{\MakeUppercase{I}}}
