### 2022-06-09 d6da061
Remove MakeUppercase
