+++ docs/latex_symbols.tex
@@ -3,14 +3,14 @@
 \DeclareMathOperator*{\argmin}{arg\,min}
 
 \newcommand{\x}{\mathbf{x}}
-\newcommand{\X}{\mathbf{\MakeUppercase{X}}}
+\newcommand{\X}{\mathbf{X}}
 \newcommand{\xx}{\mathbf{x}^{\star}}
 \newcommand{\Xx}{\mathbf{X}^{\star}}
-\newcommand{\Y}{\mathbf{\MakeUppercase{Y}}}
+\newcommand{\Y}{\mathbf{Y}}
 \newcommand{\y}{\mathbf{y}}
 \newcommand{\by}{\mathbf{y}}
 \newcommand{\chol}{\operatorname{chol}}
 \newcommand{\GP}{\mathcal{GP}}
 
-\newcommand{\Kff}{\mathbf{\MakeUppercase{K}}_{\mathbf{ff}}}
-\newcommand{\IdentityMatrix}{\mathbf{\MakeUppercase{I}}}
+\newcommand{\Kff}{\mathbf{K}_{\mathbf{ff}}}
+\newcommand{\IdentityMatrix}{\mathbf{I}}
