### 2022-06-10 debb842
Update dataclass text
