### 2022-06-13 3a68a68
Add method to combine datasets.
Use __add__ method on Dataset object to facilitates combining two datasets. This is useful for optimal design purposes.
