+++ gpjax/types.py
@@ -23,6 +23,13 @@ class Dataset:
             f"- Number of datapoints: {self.X.shape[0]}\n- Dimension: {self.X.shape[1]}"
         )
 
+    def __add__(self, other: "Dataset") -> "Dataset":
+        """Combines two datasets into one."""
+        x = jnp.concatenate((self.X, other.X))  
+        y = jnp.concatenate((self.y, other.y))  
+
+        return Dataset(X=x, y=y)
+
     @property
     def n(self) -> int:
         """The number of observations in the dataset."""
