### 2022-06-14 d125703
Fix typo in notebook + add mailmap.
