### 2022-06-14 #4455 @theorashid
Hey @dfm , I've done the first round of reviewing. It's a good package. I've commented inline whilst working through the checklist. Could I have a bit of advice on the next steps?

- There are a few points which I haven't ticked off. Let me know what you think about these: if any of them should be mentioned in an issue and dealt with
- I've made a list of things that would be "nice to have" in the package but the package is in good condition and I don't think they warrant non-acceptance. Do you think any of them worth delaying accepting the paper? From your experience, are they worth filing issues on the repo or are these all uncommon in python packages / GP packages?
