### 2022-06-14 #4455 @dfm
Thanks @theorashid! I think would be better to open an issue with those comments so that we keep a record, but that's also fine, just don't delete them when they're finished (you can just add a ~~strike-through~~ maybe?).

@thomaspinder: heads up that @theorashid's comments are currently inline above: https://github.com/openjournals/joss-reviews/issues/4455#issuecomment-1146623201
