+++ docs/nbs/gotchas.pct.py
@@ -1,0 +1,131 @@
+# -*- coding: utf-8 -*-
+# ---
+# jupyter:
+#   jupytext:
+#     cell_metadata_filter: -all
+#     custom_cell_magics: kql
+#     text_representation:
+#       extension: .py
+#       format_name: percent
+#       format_version: '1.3'
+#       jupytext_version: 1.11.2
+#   kernelspec:
+#     display_name: Python 3.9.7 ('gpjax')
+#     language: python
+#     name: python3
+# ---
+
+# %% [markdown]
+# # Common Gotchas in GPJax
+#
+# GPJax is an incredibly flexible framework due to it's low-level of abstraction. However, by providing a large degree of flexibility to users, it is possbile for pain points to be encountered when developing custom models. In this notebook we try to shed some light on these _gothcas_, providing explanations for their presence and giving appropriate solutions where applicable.
+#
+# We will not attempt to cover all Jax based edge cases in this notebook as they are covered superbly in the [Jax - The Sharp Bits](https://jax.readthedocs.io/en/latest/notebooks/Common_Gotchas_in_JAX.html) notebook.
+#
+# _This list is non-exhaustive. If you encounter your own confusion, then please consider opening [contributing](https://github.com/thomaspinder/GPJax/blob/master/CONTRIBUTING.md) a PR to this notebook._
+
+# %% [markdown]
+# ## Chex Dataclasses
+#
+# Dataclasses are supported in base Python since version 3.7. However, for a dataclass object to work with Jax, they must be a registered PyTree. This is where [Chex](https://github.com/deepmind/chex) comes in. Chex is Jax library that provides an implementation of Python's `dataclass` object that is a valid PyTree. Using dataclasses inplace of regular Python classes removes a signifacnt amount of boilerplate when defining new and thus keeps the underlying codebase more concise. However, there are a few specifics that one should be mindfule of when using dataclasses.
+#
+# ### Instantiation
+#
+# When instantiating a dataclass, a requirement is that all arguments are keyword arguments. As an example, consider the instantiation of a GP prior that is parameterised by an RBF kernel:
+# ```diff
+# ❌ prior = gpx.Prior(gpx.RBF())
+# ✅ prior = gpx.Prior(kernel=gpx.RBF())
+# ```
+# Running the former will error as the kernel argument has been supplied without an associated keyword name. Takeaway: __when instantiating GPJax object, all arguments must be named arguments__.
+#
+# ### Inheritance
+#
+# Dataclasses are just regular classes under the hood. However, there is some nuance when inheriting dataclasses from one another. Namely this boils down to the inheritance order being in reverse.
+#
+# Much like regular classes and functions, when defining a custom dataclass non-default arguments must precede the arguments with default values. However, when inheriting dataclasses, this process can be a little subtle. For example, imagine the following case where we have a generic stationary kernel
+# ```python
+# @dataclass
+# class Stationary:
+#     lengthscale: float
+#     variance: float = 1.
+# ```
+# If we now wish to inherit the `Stationary` dataclass to define a Matérn kernel, we may wish to make the smoothness parameter an additional field. However, the following code would error
+# ```python
+# @dataclass
+# class Matern(Stationary):
+#     nu: float
+# ```
+# due to the order in which dataclass fields are inherited. If we unravel this inheritance, then in this example the Python is trying to create the following class:
+# ```python
+# @dataclass
+# class Matern:
+#     lengthscale: float
+#     variance: float = 1.
+#     nu: float
+# ```
+# Hopefully now it is clear why this has errored - having the default `variance` field precede the required argument `nu` goes against Python's desgin structure. So how can this be resolved?
+#
[…]
