### 2022-06-15 thomaspinder/GPJax#76 @theorashid
JOSS review
This is for openjournals/joss-reviews#4455.

Thanks for the submission – it's a nice package with a clean codebase. Beyond some edits to the script, for which I opened a PR, there are some other points to consider (happy to discuss).

Before accepting:
- [x] Looks like jejjohnson (4,286 ++ 1,779 --) committed far more than Daniel-Dodd. The former is not on the paper, the latter is. Does their contribution warrant authorship?
- [x] the information in CONTRIBUTING.md should be linked in the repo README and the documentation. For those who aren't into software, it's especially important to visibly link to the where users can get support (looks like the discussion section to me). I hope this will increase usage of the package beyond the ML community too and into applied fields.

Not essential, but would be nice:
- ~~API which you can click through on the menu bar rather than one big dump at https://gpjax.readthedocs.io/en/latest/api.html~~
- ~~it would be good to automate linting and formatting, ideally using a pre-commit and a workflow~~
- numpyro integration. numpyro is much easier to use than tfp. I understand that perhaps distrax works better with tfp, so it would therefore be nice to see distrax in numpyro and gpjax with numpyro – it's all jax
- some more likelihoods: binomial, Poisson, Student-t etc
