+++ gpjax/variational_inference.py
@@ -2,6 +2,7 @@ import abc
 from typing import Callable, Dict
 
 import jax.numpy as jnp
+import jax.scipy as jsp
 from chex import dataclass
 from jax import vmap
 
@@ -9,9 +10,11 @@ from .gps import AbstractPosterior
 from .parameters import transform
 from .quadrature import gauss_hermite_quadrature
 from .types import Array, Dataset
-from .utils import concat_dictionaries
+from .utils import concat_dictionaries, I
 from .variational_families import AbstractVariationalFamily
-
+from .kernels import cross_covariance, diagonal, gram
+from .likelihoods import Gaussian
+from .variational_families import CollapsedVariationalGaussian
 
 @dataclass
 class AbstractVariationalInference:
@@ -111,3 +114,86 @@ class StochasticVI(AbstractVariationalInference):
         expectation = gauss_hermite_quadrature(log_prob, mean, variance, y=y)
 
         return expectation
+
+
+@dataclass
+class CollapsedVI(AbstractVariationalInference):
+    """Collapsed variational inference for a sparse Gaussian process regression model. 
+    The key reference is Titsias, (2009) - Variational Learning of Inducing Variables in Sparse Gaussian Processes."""
+
+    def __post_init__(self):
+        self.prior = self.posterior.prior
+        self.likelihood = self.posterior.likelihood
+        self.num_inducing = self.variational_family.num_inducing
+        
+        if not isinstance(self.likelihood, Gaussian):
+            raise TypeError("Likelihood must be Gaussian.")
+            
+        if not isinstance(self.variational_family, CollapsedVariationalGaussian):
+            raise TypeError("Variational family must be CollapsedVariationalGaussian.")
+
+    def elbo(
+        self, train_data: Dataset, transformations: Dict, negative: bool = False
+    ) -> Callable[[Array], Array]:
+        """Compute the evidence lower bound under this model. In short, this requires evaluating the expectation of the model's log-likelihood under the variational approximation. To this, we sum the KL divergence from the variational posterior to the prior. When batching occurs, the result is scaled by the batch size relative to the full dataset size.
+
+        Args:
+            train_data (Dataset): The training data for which we should maximise the ELBO with respect to.
+            transformations (Dict): The transformation set that unconstrains each parameter.
+            negative (bool, optional): Whether or not the resultant elbo function should be negative. For gradient descent where we minimise our objective function this argument should be true as minimisation of the negative corresponds to maximisation of the ELBO. Defaults to False.
+
+        Returns:
+            Callable[[Dict, Dataset], Array]: A callable function that accepts a current parameter estimate for which gradients should be computed.
+        """
+        constant = jnp.array(-1.0) if negative else jnp.array(1.0)
+        
+        x, y, n = train_data.X, train_data.y, train_data.n
+        
+        m = self.num_inducing
+       
+
+        def elbo_fn(params: Dict) -> Array:
+            params = transform(params, transformations)
+            noise = params["likelihood"]["obs_noise"]
+            z = params["variational_family"]["inducing_inputs"]
+            Kzz = gram(self.prior.kernel, z, params["kernel"])
[…]
