### 2022-06-15 537797b
Add skeleton notebook.
