### 2022-06-15 1bc6d8b
Add skeleton notebook.
