### 2022-06-15 021a80e
Laplace approximation example.
