+++ docs/nbs/classification.pct.py
@@ -26,6 +26,9 @@ import matplotlib.pyplot as plt
 import blackjax
 import gpjax as gpx
 import optax as ox
+import distrax as dx
+from gpjax.utils import I
+import jax.scipy as jsp
 
 key = jr.PRNGKey(123)
 
@@ -72,15 +75,145 @@ mll = jax.jit(posterior.marginal_log_likelihood(D, constrainer, negative=True))
 # We can obtain a MAP estimate by optimising the marginal log-likelihood with Obtax's optimisers.
 # %%
 opt = ox.adam(learning_rate=0.01)
-map_estimate = gpx.fit(
+unconstrained_params = gpx.fit(
     mll,
     params,
     trainable,
     opt,
     n_iters=500,
 )
+
+negative_Hessian = jax.jacfwd(jax.jacrev(mll))(unconstrained_params)["latent"]["latent"][:,0,:,0]
+
+map_estimate = gpx.transform(unconstrained_params, constrainer)
 # %% [markdown]
-# However, as a point estimate, MAP estimation is severely limited for uncertainty quantification, providing only a single piece of information about the posterior. On the other hand, through approximate sampling, MCMC methods allow us to learn all information about the posterior distribution.
+# From which we can make predictions at novel inputs, as illustrated below.
+# %%
+latent_dist = posterior(D, map_estimate)(xtest)
+
+predictive_dist = likelihood(latent_dist, map_estimate)
+
+predictive_mean = predictive_dist.mean()
+predictive_std = predictive_dist.stddev()
+
+fig, ax = plt.subplots(figsize=(12, 5))
+ax.plot(x, y, "o", label="Observations", color="tab:red")
+ax.plot(xtest, predictive_mean, label="Predictive mean", color="tab:blue")
+ax.fill_between(
+    xtest.squeeze(),
+    predictive_mean - predictive_std,
+    predictive_mean + predictive_std,
+    alpha=0.2,
+    color="tab:blue",
+    label='Two sigma',
+)
+ax.plot(xtest, predictive_mean - predictive_std, color="tab:blue", linestyle="--", linewidth=1)
+ax.plot(xtest, predictive_mean + predictive_std, color="tab:blue", linestyle="--", linewidth=1)
+
+ax.legend()
+
+# %% [markdown]
+# However, as a point estimate, MAP estimation is severely limited for uncertainty quantification, providing only a single piece of information about the posterior.
+# %% [markdown]
+# ## Laplace approximation
+# The Laplace approximation improves uncertainty quantification by incorporating curvature induced by the marginal log-likelihood's Hessian to construct an approximate Gaussian distribution centered on the MAP estimate.
+# Since the negative Hessian is positive definite, we can use the Cholesky decomposition to obtain the covariance matrix of the Laplace approximation at the datapoints below.
+# %%
+f_map_estimate = posterior(D, map_estimate)(x).mean()
+
+jitter = 1e-6
+
+# LLᵀ = H
+L = jnp.linalg.cholesky(negative_Hessian + I(D.n) * jitter)
+
+# H⁻¹ = H⁻¹ I = (LLᵀ)⁻¹ I = L⁻ᵀL⁻¹ I
+L_inv = jsp.linalg.solve_triangular(L, I(D.n), lower=True)
+H_inv = jsp.linalg.solve_triangular(L.T, L_inv, lower=False)
+
[…]
