### 2022-06-15 2aa6a38
Delete gotchas.pct.py
Didn't mean to include within this PR.
