### 2022-06-15 2d3d8e4
Remove gotchas reference from index.rst
