### 2022-06-15 thomaspinder/GPJax#76 @thomaspinder
Hey @theorashid, thanks for giving us such a helpful set of comments. I'll address them in order below.

> API which you can click through on the menu bar rather than one big dump at https://gpjax.readthedocs.io/en/latest/api.html

The API page has a right hand menu which is structured as `module > class `. May I ask what benefit you feel having a menu bar would give over the current right-hand menu?

> numpyro integration. numpyro is much easier to use than tfp. I understand that perhaps distrax works better with tfp, so it would therefore be nice to see distrax in numpyro and gpjax with numpyro – it's all jax

I do not have a great deal of experience with NumPyro. We chose to offload MCMC sampling into BlackJax and optimisation to Optax as they were both quite "research focussed" packages which aligned with the principles in GPJax. I'm curious, would you mind expanding on the advantages that NumPyro integration would give in comparison to Optax and BlackJax?

> the information in CONTRIBUTING.md should be linked in the repo README and the documentation. For those who aren't into software, it's especially important to visibly link to the where users can get support (looks like the discussion section to me). I hope this will increase usage of the package beyond the ML community too and into applied fields.

Good point - this has been resolved [here](https://github.com/thomaspinder/GPJax/pull/77).

> it would be good to automate linting and formatting, ideally using a pre-commit and a workflow

Cool suggestion. A pre-commit automation that uses the black formatter has been added [here](https://github.com/thomaspinder/GPJax/pull/77).

> some more likelihoods: binomial, Poisson, Student-t etc

This is on our radar. As is a guide detailing custom implementation. We would like to give it a bit more thought though as our end goal is to have an API such that one can simply inherit a Distrax distribution and define one or two methods.

> Looks like jejjohnson (4,286 ++ 1,779 --) committed far more than Daniel-Dodd. The former is not on the paper, the latter is. Does their contribution warrant authorship?

A major reason for jejjohnson's large number of committed lines of code is that we used `.ipynb` files for notebooks originally, not `.pct.py` files. Therefore, each time a notebook was run, the corresponding commit would have hundreds of lines corresponding to the notebook's output. This also explains why I have such a large number of committed lines. jejjohnson contributed a series of functions that enabled integration with NumPyro. However, this was in a very early version of GPJax (last commit was March 2021) and the package has evolved significantly since then. Many of the commits made by Daniel-Dodd have been attributed to the user Daniel Dodd due to a changing email address. The true number of lines committed by Daniel-Dodd is actually ++2813 --2012. This line count does not truly reflect the contributions made by Daniel-Dodd though as he is the main contributor the sparse framework within GPJax and was heavily influential in curating the example notebooks within the documentation. I am very happy to discuss the criteria for authorship more if you disagree.
