### 2022-06-15 thomaspinder/GPJax#76 @theorashid
Many of those points are fair and I've ticked them off. I don't have much experience with pre-commit but make sure it runs for every PR, and maybe add it to the CONTRIBUTING.md. Glad it was easy enough to implement.

> May I ask what benefit you feel having a menu bar would give over the current right-hand menu?

Ah okay, I didn't notice the toggle for the menu in the top right. I see it now and this is all good. If there's a way to set it as open by default, that would be great.

> I do not have a great deal of experience with NumPyro. We chose to offload MCMC sampling into BlackJax and optimisation to Optax as they were both quite "research focussed" packages which aligned with the principles in GPJax. I'm curious, would you mind expanding on the advantages that NumPyro integration would give in comparison to Optax and BlackJax?

It's more for the cases when you want to use GPs as part of a larger model. In spatial modelling, perhaps you only want a Gaussian process over space in a larger model with several covariates added as $\beta*covariate$. Although, now searching for this, I'm not sure how many examples there are of these beyond forum posts or custom GP implementations. Perhaps the editor @dfm has better ideas with this and whether it is achievable.

As for NumPyro specifically, I just prefer using it for model building to tfp. And it's written in jax. So, taking from [your tfp example](https://gpjax.readthedocs.io/en/latest/nbs/tfp_integration.html), it would be nice if you could replace
```diff
-priors["kernel"]["lengthscale"] = tfd.Gamma(concentration=jnp.array(1.0), rate=jnp.array(1.0))
+priors["kernel"]["lengthscale"] = numpyro.sample("rho", numpyro.dist.Gamma(1.0, 10.0))
```
and equally the numpyro mcmc interface is much simpler imo than tfp with it's confusing `trace_fn` argument, where you can just use `mcmc.run()`. This would encourage users from applied fields with less experience of mcmc.

Personally, I would always use blackjax for mcmc over tfp, especially with their new API.

> We would like to give it a bit more thought though as our end goal is to have an API such that one can simply inherit a Distrax distribution and define one or two methods.

I guess you could look towards how the tingp library [has done this with numpyro](https://tinygp.readthedocs.io/en/stable/tutorials/likelihoods.html).
