### 2022-06-15 35173f8
Add notion of "moments"
Add notion of "moments" as a moment parameterisation for a variational Gaussian.
