### 2022-06-15 218ebf4
Add AbstractVariationalGaussian class.
