### 2022-06-15 642cc73
Update test_variational_families.py
