### 2022-06-15 347be8b
Update natural_gradients.py
