+++ gpjax/natural_gradients.py
@@ -1,8 +1,9 @@
-from cmath import exp
 import typing as tp
 import jax.numpy as jnp
 import jax.scipy as jsp
 from jax import jacobian
+import distrax as dx
+from jax import lax
 
 from .config import get_defaults
 from .variational_families import AbstractVariationalFamily, ExpectationVariationalGaussian
@@ -14,21 +15,19 @@ from .parameters import build_identity, transform
 
 DEFAULT_JITTER = get_defaults()["jitter"]
 
-# CURRENTLY THIS FILE IS A FIRST SKETCH OF NATURAL GRADIENTS in GPJax.
 
-# Below is correct, but it might be better to pass in params (i.e., all svgp params) and return a dictionary that gives svgp params
-def natural_to_expectation(natural_params: dict, jitter: float = DEFAULT_JITTER):
+def natural_to_expectation(natural_moments: dict, jitter: float = DEFAULT_JITTER):
     """
     Converts natural parameters to expectation parameters.
     Args:
-        natural_params: A dictionary of natural parameters.
-        jitter: A small value to prevent numerical instability.
+        natural_moments: A dictionary of natural parameters.
+        jitter (float): A small value to prevent numerical instability.
     Returns:
-        A dictionary of expectation parameters.
+        tp.Dict: A dictionary of Gaussian moments under the expectation parameterisation.
     """
     
-    natural_matrix = natural_params["natural_matrix"]
-    natural_vector = natural_params["natural_vector"]
+    natural_matrix = natural_moments["natural_matrix"]
+    natural_vector = natural_moments["natural_vector"]
     m = natural_vector.shape[0]
     
     # S⁻¹ = -2θ₂
@@ -56,63 +55,76 @@ def natural_to_expectation(natural_params: dict, jitter: float = DEFAULT_JITTER)
     return {"expectation_vector": expectation_vector, "expectation_matrix": expectation_matrix}
 
 
-# This is a function that you create before training. This can be used to get the elbo for the nexpectation parameterisation.
-# Here it is assumed that the parameters have already been transformed prior to being passed to the returned function.
-def get_expectation_elbo(posterior: AbstractPosterior,
+def _expectation_elbo(posterior: AbstractPosterior,
             variational_family: AbstractVariationalFamily,
             train_data: Dataset,
             ):
     """
-    Computes evidence lower bound (ELBO) for the expectation parameterisation.
+    Construct evidence lower bound (ELBO) for varational Gaussian under the expectation parameterisation.
     Args:
         posterior: An instance of AbstractPosterior.
         variational_family: An instance of AbstractVariationalFamily.
     Returns:
         Callable: A function that computes ELBO.
     """
-    q = variational_family
-    expectaction_q = ExpectationVariationalGaussian(prior=q.prior, inducing_inputs = q.inducing_inputs)
-    svgp = StochasticVI(posterior=posterior, variational_family=expectaction_q)
-    transformations = build_identity(svgp.params)
+    evg = ExpectationVariationalGaussian(prior= variational_family.prior, 
+                                        inducing_inputs = variational_family.inducing_inputs,
+                                        )
+    svgp = StochasticVI(posterior=posterior, variational_family=evg)
 
-    return svgp.elbo(train_data, transformations)
+    return svgp.elbo(train_data, build_identity(svgp.params))
 
 
-def natural_gradients(params: dict,
-            transformations: dict,
-            expectation_elbo: tp.Callable,
-            nat_to_xi: tp.Callable, 
-            xi_to_nat: tp.Callable, 
-            batch,
+def natural_gradients(
+    posterior: AbstractPosterior,
+    variational_family: AbstractVariationalFamily,
[…]
