### 2022-06-15 37bedd6
Update natural_gradients.py
