+++ paper.bib
@@ -26,7 +26,7 @@
 
 
 @article{gardner2018gpytorch,
-  title={{Gpytorch}: Blackbox matrix-matrix {G}aussian process inference with {GPU} acceleration},
+  title={{G{p}y{t}orch}: Blackbox matrix-matrix {G}aussian process inference with {GPU} acceleration},
   author={Gardner, Jacob and Pleiss, Geoff and Weinberger, Kilian Q and Bindel, David and Wilson, Andrew G},
   journal={Advances in neural information processing systems},
   volume={31},
@@ -140,8 +140,8 @@
   title =    {{GPy}: A {Gaussian} process framework in python},
   url = {http://github.com/SheffieldML/GPy},
   year = {2012},
-  publisher={Github},
-  journal={Github repository},
+  publisher={Git{H}ub},
+  journal={Git{H}ub repository},
 }
 
 @Misc{blackjax2021,
@@ -149,8 +149,8 @@
   title =    {{BlackJAX}; a sampling library designed for ease of use, speed and modularity. },
   url = {https://github.com/blackjax-devs/blackjax/},
   year = {2021},
-  publisher={Github},
-  journal={Github repository},
+  publisher={Git{H}ub},
+  journal={Git{H}ub repository},
 }
 
 @Misc{stheno2022tebbutt,
@@ -158,8 +158,8 @@
   author = {Tebbutt W, Bruinsma W, Turner RE},
   url={https://github.com/JuliaGaussianProcesses/Stheno.jl},
   year = {2022},
-  publisher={Github},
-  journal={Github repository},
+  publisher={Git{H}ub},
+  journal={Git{H}ub repository},
 }
 
 @Misc{dfm2021tinygp,
@@ -167,8 +167,8 @@
   author = {Dan Foreman-Mackey},
   url={https://github.com/dfm/tinygp},
   year = {2021},
-  publisher={Github},
-  journal={Github repository},
+  publisher={Git{H}ub},
+  journal={Git{H}ub repository},
 }
 
 @Misc{objax2020github,
@@ -242,7 +242,7 @@
 
 
 @article{mockus1978application,
-  title={The application of Bayesian methods for seeking the extremum},
+  title={The application of {B}ayesian methods for seeking the extremum},
   author={Mockus, Jonas and Tiesis, Vytautas and Zilinskas, Antanas},
   journal={Towards global optimization},
   volume={2},
