### 2022-06-15 thomaspinder/GPJax#76 @theorashid
Sorry if it wasn't clear: only the two points on the checklist at the top and the paper were essential to accepting the submission. In my view, the codebase is more than ready to be accepted – I just need to read and edit the latest version of the manuscript. The numpyro stuff and menu are **not** essential at all. Obviously there's another reviewer but great job.

> Yes, very cool. I really like this idea. If you were interested, I'd be very interested to talk more about the logistics of doing this with a nice real-world example in GPJax.

I'd be really interested – particularly for my own work. I think a spatiotemporal problem would be a nice way to involve that community with your package. A good example dataset would be the [Hungary Chickenpox dataset](https://pytorch-geometric-temporal.readthedocs.io/en/latest/modules/dataset.html#module-torch_geometric_temporal.dataset.chickenpox) . I would like to do this with a graph GP over space kroneckered with a GP over time (see [Fast Hierarchical GPs](https://sethrf.com/files/fast-hierarchical-GPs.pdf)) if that's possible. It's also a Poisson likelihood and we could try and add a covariate in for a larger model. Perhaps when you're more free, we could sit down for a chat about implementing this – many in my spatiotemproal epidemiology crowd would be interested.

Also, leave this issue open for now and I'll put any comments I have about the paper here before closing it.
