### 2022-06-16 #4455 @thomaspinder
Updating the pdf as we have made capitalisation of paper titles consistent now in our references e.g., "Gaussian" -> "{G}aussian".
