+++ docs/requirements.txt
@@ -12,7 +12,7 @@ ipython
 ipykernel
 watermark
 sphinxext-opengraph
-blackjax
+blackjax==0.7.0
 dm-haiku
 ipywidgets
 # Install GPJax istself
