### 2022-06-16 thomaspinder/GPJax#76 @thomaspinder
> I'd be really interested – particularly for my own work.

Great. Very interested to explore this more, particularly in relation to the graph component with a temporal component. I have opened a discussion thread #80 for us to continue publicly discussing this idea further.

> Also, leave this issue open for now and I'll put any comments I have about the paper here before closing it.

Sure! Look forward to anymore thoughts you have on the package.
