+++ docs/nbs/regression.pct.py
@@ -1,209 +1,0 @@
-# ---
-# jupyter:
-#   jupytext:
-#     custom_cell_magics: kql
-#     text_representation:
-#       extension: .py
-#       format_name: percent
-#       format_version: '1.3'
-#       jupytext_version: 1.11.2
-#   kernelspec:
-#     display_name: Python 3.9.7 ('gpjax')
-#     language: python
-#     name: python3
-# ---
-
-# %% [markdown] pycharm={"name": "#%% md\n"}
-# # Regression
-#
-# In this notebook we demonstate how to fit a Gaussian process regression model.
-
-# %%
-from pprint import PrettyPrinter
-
-import jax.numpy as jnp
-import jax.random as jr
-from jax import jit
-
-import matplotlib.pyplot as plt
-import gpjax as gpx
-import optax as ox
-
-pp = PrettyPrinter(indent=4)
-key = jr.PRNGKey(123)
-# %% [markdown]
-# ## Dataset
-#
-# With the necessary modules imported, we simulate a dataset $\mathcal{D} = (\boldsymbol{x}, \boldsymbol{y}) = \{(x_i, y_i)\}_{i=1}^{100}$ with inputs $\boldsymbol{x}$ sampled uniformly on $(-3., 3)$ and corresponding independent noisy outputs
-#
-# $$\boldsymbol{y} \sim \mathcal{N} \left(\sin(4\boldsymbol{x}) + \cos(2 \boldsymbol{x}), \textbf{I} * 0.3^2 \right).$$
-#
-# We store our data $\mathcal{D}$ as a GPJax `Dataset` and create test inputs and labels for later.
-# %%
-n = 100
-noise = 0.3
-
-x = jr.uniform(key=key, minval=-3.0, maxval=3.0, shape=(n,)).sort().reshape(-1, 1)
-f = lambda x: jnp.sin(4 * x) + jnp.cos(2 * x)
-signal = f(x)
-y = signal + jr.normal(key, shape=signal.shape) * noise
-
-D = gpx.Dataset(X=x, y=y)
-
-xtest = jnp.linspace(-3.5, 3.5, 500).reshape(-1, 1)
-ytest = f(xtest)
-
-# %% [markdown]
-# To better understand what we have simulated, we plot both the underlying latent function and the observed data that is subject to Gaussian noise.
-# %%
-fig, ax = plt.subplots(figsize=(10, 5))
-ax.plot(xtest, ytest, label="Latent function")
-ax.plot(x, y, "o", label="Observations")
-ax.legend(loc="best")
-# %% [markdown]
-# Our aim in this tutorial will be to reconstruct the latent function from our noisy observations $\mathcal{D}$ via Gaussian process regression. We begin by defining a Gaussian process prior in the next section.
-
-
-# %% [markdown]
-# ## Defining the prior
-#
-# A zero-mean Gaussian process (GP) places a prior distribution over real-valued functions $f(\cdot)$ where $f(\boldsymbol{x}) \sim \mathcal{N}(0, \mathbf{K}_{\boldsymbol{x}\boldsymbol{x}})$ for any finite collection of inputs $\boldsymbol{x}$. 
-# Here $\mathbf{K}_{\boldsymbol{x}\boldsymbol{x}}$ is the Gram matrix generated by a user-specified symmetric, non-negative definite kernel function $k(\cdot, \cdot')$ with $[\mathbf{K}_{\boldsymbol{x}\boldsymbol{x}}]_{i, j} = k(x_i, x_j)$. 
-# The choice of kernel function is critical as, among other things, it governs the smoothness of the outputs that our GP can generate. 
[…]
