+++ docs/nbs/kernels.pct.py
@@ -1,293 +1,0 @@
-# -*- coding: utf-8 -*-
-# ---
-# jupyter:
-#   jupytext:
-#     custom_cell_magics: kql
-#     text_representation:
-#       extension: .py
-#       format_name: percent
-#       format_version: '1.3'
-#       jupytext_version: 1.11.2
-#   kernelspec:
-#     display_name: Python 3.9.7 ('gpjax')
-#     language: python
-#     name: python3
-# ---
-
-# %% [markdown]
-# # Kernel Guide
-#
-# In this guide, we introduce the kernels available in GPJax and demonstrate how to create custom ones.
-
-# %%
-import jax.numpy as jnp
-import jax.random as jr
-import matplotlib.pyplot as plt
-import tensorflow_probability.substrates.jax.bijectors as tfb
-from jax import jit
-import jax
-from optax import adam
-import distrax as dx
-
-import gpjax as gpx
-
-key = jr.PRNGKey(123)
-
-# %% [markdown]
-# ## Supported Kernels
-#
-# The following kernels are natively supported in GPJax.
-#
-# * Matérn 1/2, 3/2 and 5/2.
-# * RBF (or squared exponential).
-# * Polynomial.
-# * [Graph kernels](https://gpjax.readthedocs.io/en/latest/nbs/graph_kernels.html).
-#
-# While the syntax is consistent, each kernel’s type influences the characteristics of the sample paths drawn. We visualise this below with 10 function draws per kernel.
-
-# %%
-kernels = [
-    gpx.Matern12(),
-    gpx.Matern32(),
-    gpx.Matern52(),
-    gpx.RBF(),
-    gpx.Polynomial(degree=1),
-    gpx.Polynomial(degree=2),
-]
-fig, axes = plt.subplots(ncols=3, nrows=2, figsize=(20, 10))
-
-x = jnp.linspace(-3.0, 3.0, num=200).reshape(-1, 1)
-
-
-for k, ax in zip(kernels, axes.ravel()):
-    prior = gpx.Prior(kernel=k)
-    params, _, _, _ = gpx.initialise(prior)
-    rv = prior(params)(x)
-    y = rv.sample(sample_shape=10, seed=key)
-
-    ax.plot(x, y.T, alpha=0.7)
-    ax.set_title(k.name)
-
-# %% [markdown]
-# ### Active dimensions
-#
-# By default, kernels operate over every dimension of the supplied inputs. In some use cases, it is desirable to restrict kernels to specific dimensions of the input data. We can achieve this by the `active dims` argument, which determines which input index values the kernel evaluates.
-#
-# To see this, consider the following 5-dimensional dataset for which we would like our RBF kernel to act on the first, second and fourth dimensions.
-
-# %%
-slice_kernel = gpx.RBF(active_dims=[0, 1, 3])
-
-# %% [markdown]
-#
-# The resulting kernel has one length-scale parameter per input dimension --- an ARD kernel.
-
-# %%
-print(f"ARD: {slice_kernel.ard}")
-print(f"Lengthscales: {slice_kernel.params['lengthscale']}")
-
-# %% [markdown]
-# We'll now simulate some data and evaluate the kernel on the previously selected input dimensions.
-
-# %%
-x_matrix = jr.normal(key, shape=(50, 5))
-K = gpx.kernels.gram(slice_kernel, x_matrix, slice_kernel.params)
-print(K.shape)
-
-# %% [markdown]
-# ## Kernel combinations
-#
[…]
