### 2022-06-16 02f068b
Minimal working natural gradient functions for NAT PARAMETERISATION.
