+++ gpjax/natural_gradients.py
@@ -1,23 +1,23 @@
+from copy import deepcopy
+from multiprocessing.dummy import Array
 import typing as tp
 import jax.numpy as jnp
 import jax.scipy as jsp
-from jax import jacobian
 import distrax as dx
-from jax import lax
-import jax
+from jax import lax, value_and_grad
 
 from .config import get_defaults
-from .variational_families import AbstractVariationalFamily, ExpectationVariationalGaussian
+from .variational_families import AbstractVariationalFamily, ExpectationVariationalGaussian, NaturalVariationalGaussian
 from .variational_inference import StochasticVI
 from .utils import I
 from .gps import AbstractPosterior
 from .types import Dataset
-from .parameters import build_identity, transform
+from .parameters import Identity, build_identity, transform, build_trainables_false, build_trainables_true, trainable_params
 
 DEFAULT_JITTER = get_defaults()["jitter"]
 
 
-def natural_to_expectation(natural_moments: dict, jitter: float = DEFAULT_JITTER):
+def natural_to_expectation(natural_moments: dict, jitter: float = DEFAULT_JITTER) -> dict:
     """
     Converts natural parameters to expectation parameters.
     Args:
@@ -59,7 +59,7 @@ def natural_to_expectation(natural_moments: dict, jitter: float = DEFAULT_JITTER
 def _expectation_elbo(posterior: AbstractPosterior,
             variational_family: AbstractVariationalFamily,
             train_data: Dataset,
-            ):
+            ) -> tp.Callable[[dict, Dataset], float]:
     """
     Construct evidence lower bound (ELBO) for varational Gaussian under the expectation parameterisation.
     Args:
@@ -76,61 +76,150 @@ def _expectation_elbo(posterior: AbstractPosterior,
     return svgp.elbo(train_data, build_identity(svgp.params))
 
 
-
-#DOES NOT WORK YET.
-def natural_gradients(
-    posterior: AbstractPosterior,
-    variational_family: AbstractVariationalFamily,
-    train_data: Dataset,
-    params: dict,
-    transformations: dict,
-    batch,
-    nat_to_moments: tp.Optional[dx.Bijector] = Identity,
-    ) -> dict:
+def _stop_gradients_nonmoments(params: tp.Dict) -> tp.Dict:
     """
-    Computes natural gradients for a variational family.
+    Stops gradients for non-moment parameters.
     Args:
-        posterior (AbstractPosterior): An instance of AbstractPosterior.
-        variational_family(AbstractVariationalFamily): An instance of AbstractVariationalFamily.
-        train_data (Dataset): Training Dataset.
-        params (tp.Dict): A dictionary of model parameters.
-        transformations (tp.Dict): A dictionary of parameter transformations.
-        nat_to_moments (dx.Bijector): A bijector between natural and the chosen parameterisations of the Gaussian variational moments.
+        params: A dictionary of parameters.
     Returns:
-        tp.Dict: Dictionary of natural gradients.
+        tp.Dict: A dictionary of parameters with stopped gradients.
     """
-    # Transform the parameters.
-    params = transform(params, transformations)
+    trainables = build_trainables_false(params)
+    moment_trainables = build_trainables_true(params["variational_family"]["moments"])
+    trainables["variational_family"]["moments"] = moment_trainables
+    params = trainable_params(params, trainables)
+    return params
 
-    # Get moments and stop gradients for non-moment parameters.
-    moments = params["variational_family"]["moments"]
+def _stop_gradients_moments(params: tp.Dict) -> tp.Dict:
+    """
+    Stops gradients for moment parameters.
+    Args:
+        params: A dictionary of parameters.
+    Returns:
+        tp.Dict: A dictionary of parameters with stopped gradients.
+    """
[…]
