+++ tests/test_natural_gradients.py
@@ -1,0 +1,136 @@
+import pytest
+import jax
+import jax.numpy as jnp
+from gpjax.natural_gradients import natural_to_expectation, _stop_gradients_nonmoments, _stop_gradients_moments, _expectation_elbo, natural_gradients
+import typing as tp
+import gpjax as gpx
+import jax.random as jr
+from gpjax.parameters import recursive_items
+import jax.numpy as jnp
+import jax.random as jr
+import matplotlib.pyplot as plt
+from jax import jit
+import optax as ox
+
+import gpjax as gpx
+import tensorflow as tf
+
+tf.random.set_seed(42)
+key = jr.PRNGKey(123)
+
+@pytest.mark.parametrize("dim", [1, 2, 3])
+def test_natural_to_expectation(dim):
+    """
+    Converts natural parameters to expectation parameters.
+    Args:
+        natural_moments: A dictionary of natural parameters.
+        jitter (float): A small value to prevent numerical instability.
+    Returns:
+        tp.Dict: A dictionary of Gaussian moments under the expectation parameterisation.
+    """
+
+    natural_matrix = -.5 * jnp.eye(dim)
+    natural_vector = jnp.zeros((dim, 1))
+    
+    natural_moments = {"natural_matrix": natural_matrix, "natural_vector": natural_vector}
+
+    expectation_moments = natural_to_expectation(natural_moments, jitter=1e-6)
+
+    assert "expectation_vector" in expectation_moments.keys()
+    assert "expectation_matrix" in expectation_moments.keys()
+    assert expectation_moments["expectation_vector"].shape == natural_moments["natural_vector"].shape
+    assert expectation_moments["expectation_matrix"].shape == natural_moments["natural_matrix"].shape
+
+
+def get_data_and_gp(n_datapoints):
+    x = jnp.linspace(-5.0, 5.0, n_datapoints).reshape(-1, 1)
+    y = jnp.sin(x) + jr.normal(key=jr.PRNGKey(123), shape=x.shape) * 0.1
+    D = gpx.Dataset(X=x, y=y)
+
+    p = gpx.Prior(kernel=gpx.RBF())
+    lik = gpx.Gaussian(num_datapoints=n_datapoints)
+    post = p * lik
+    return D, post, p
+
+@pytest.mark.parametrize("jit_fns", [True, False])
+def test_expectation_elbo(jit_fns):
+    """
+    Tests the expectation ELBO.
+    """
+    D, posterior, prior = get_data_and_gp(10)
+    
+    z = jnp.linspace(-5.0, 5.0, 5).reshape(-1, 1)
+    variational_family = gpx.variational_families.ExpectationVariationalGaussian(prior = prior, inducing_inputs=z)
+
+    svgp = gpx.StochasticVI(posterior=posterior, variational_family=variational_family)
+
+    params, _, constrainer, unconstrainer = gpx.initialise(svgp)
+
+    expectation_elbo = _expectation_elbo(posterior, variational_family, D)
+
+    if jit_fns:
+        elbo_fn = jax.jit(expectation_elbo)
+    else:
+        elbo_fn = expectation_elbo
+
+    assert isinstance(elbo_fn, tp.Callable)
+    elbo_value = elbo_fn(params, D)
+    assert isinstance(elbo_value, jnp.ndarray)
+
+    # Test gradients
+    grads = jax.grad(elbo_fn, argnums=0)(params, D)
+    assert isinstance(grads, tp.Dict)
+    assert len(grads) == len(params)
+
+
+
+# def test_stop_gradients_nonmoments():
+#     pass
+    
+
[…]
