### 2022-06-16 25a7893
Update natural_gradients.py
