### 2022-06-17 a5499d0
Update natural_gradients.py
Fix negative, and add sketch optimisation loop.
