+++ gpjax/variational_inference.py
@@ -7,14 +7,17 @@ from chex import dataclass
 from jax import vmap
 
 from .gps import AbstractPosterior
+from .kernels import cross_covariance, diagonal, gram
+from .likelihoods import Gaussian
 from .parameters import transform
 from .quadrature import gauss_hermite_quadrature
 from .types import Array, Dataset
-from .utils import concat_dictionaries, I
-from .variational_families import AbstractVariationalFamily
-from .kernels import cross_covariance, diagonal, gram
-from .likelihoods import Gaussian
-from .variational_families import CollapsedVariationalGaussian
+from .utils import I, concat_dictionaries
+from .variational_families import (
+    AbstractVariationalFamily,
+    CollapsedVariationalGaussian,
+)
+
 
 @dataclass
 class AbstractVariationalInference:
@@ -31,7 +34,7 @@ class AbstractVariationalInference:
     def params(self) -> Dict:
         """Construct the parameter set used within the variational scheme adopted."""
         hyperparams = concat_dictionaries(
-            {"likelihood": self.posterior.likelihood.params}, 
+            {"likelihood": self.posterior.likelihood.params},
             self.variational_family.params,
         )
         return hyperparams
@@ -104,11 +107,15 @@ class StochasticVI(AbstractVariationalInference):
 
         # q(f(x))
         predictive_dist = vmap(self.variational_family.predict(params))(x[:, None])
-        mean = predictive_dist.mean().val.reshape(-1,1)
-        variance = predictive_dist.variance().val.reshape(-1,1)
+        mean = predictive_dist.mean().val.reshape(-1, 1)
+        variance = predictive_dist.variance().val.reshape(-1, 1)
 
         # log(p(y|f(x)))
-        log_prob = vmap(lambda f, y: self.likelihood.link_function(f, params["likelihood"]).log_prob(y))
+        log_prob = vmap(
+            lambda f, y: self.likelihood.link_function(
+                f, params["likelihood"]
+            ).log_prob(y)
+        )
 
         # ≈ ∫[log(p(y|f(x))) q(f(x))] df(x)
         expectation = gauss_hermite_quadrature(log_prob, mean, variance, y=y)
@@ -118,17 +125,17 @@ class StochasticVI(AbstractVariationalInference):
 
 @dataclass
 class CollapsedVI(AbstractVariationalInference):
-    """Collapsed variational inference for a sparse Gaussian process regression model. 
+    """Collapsed variational inference for a sparse Gaussian process regression model.
     The key reference is Titsias, (2009) - Variational Learning of Inducing Variables in Sparse Gaussian Processes."""
 
     def __post_init__(self):
         self.prior = self.posterior.prior
         self.likelihood = self.posterior.likelihood
         self.num_inducing = self.variational_family.num_inducing
-        
+
         if not isinstance(self.likelihood, Gaussian):
             raise TypeError("Likelihood must be Gaussian.")
-            
+
         if not isinstance(self.variational_family, CollapsedVariationalGaussian):
             raise TypeError("Variational family must be CollapsedVariationalGaussian.")
 
@@ -146,11 +153,10 @@ class CollapsedVI(AbstractVariationalInference):
[…]
