### 2022-06-20 thomaspinder/GPJax#76 @theorashid
Paper looks in good shape. Notes:
- The two R packages you cited are very niche and I think you can drop them. I still can't believe the lack of GP software for R. Having done some digging, I found a paper for [GPfit](https://www.jstatsoft.org/htaccess.php?volume=064&type=i&issue=12&filename=paper) in R, which cites two older packages [tgp](https://bobby.gramacy.com/r_packages/tgp/) and [mlegp](https://www.rdocumentation.org/packages/mlegp/versions/3.1.9). GPfit hasn't been updated for 3 years, and the other two look niche and not widely used. I asked my supervisor and he said [GPstuff](https://github.com/gpstuff-dev/gpstuff) has a port for R, and people also use INLA and mgcv. I would probably say "...include GPML and GPstuff in Matlab, the latter is also available in R, GaussianProcess.jl..." (word however you want, how I did it there is quite clunky). Or maybe ask some R users what they do
- Stheno also has a [python version with jax functionality](https://wesselb.github.io/2021/01/19/linear-models-with-stheno-and-jax.html). This is worth mentioning in the bayesnewton/tinygp section, because Stheno _does_ have inducing point functionality. I haven't used Stheno much before, and I'm sure there are plenty of things GPJax does better.
- Up to you who you acknowledge – perhaps Erik Bodin's 2 lines were life-changing. But you can do what you want in this section
