### 2022-06-20 db9ce1d
Update variational_inference.py
