+++ paper.md
@@ -40,9 +40,9 @@ The intimacy between `GPJax` and the underlying maths also makes `GPJax` an exce
 
 # Wider Software Ecosystem
 
-From both an applied and methodological perspective, GPs are widely employed in the statistics and machine learning communities. High-quality software packages that promote GP modelling are accountable for much of their success. Within the Python community, the three most popular packages for GP modelling are GPFlow [@matthews2017gpflow], GPyTorch [@gardner2018gpytorch], and GPy [@gpy2014]. Despite each of these packages being indispensable tools for the community, none of them support integration with a Jax-based workflow. Both the BayesNewton [@wilkinson2021bayesnewton] and TinyGP [@dfm2021tinygp] packages are built on top of Jax. However, BayesNewton is designed on top of ObJax [@objax2020github], meaning that integration with the wider suite of packages in the Jax ecosystem can be challenging. Meanwhile TinyGP offers excellent integration with inference frameworks such as NumPyro [@phan2019composable] but does not yet support inducing points frameworks (e.g., [@hensman2013gaussian]). `GPJax` exists to resolve this issues. Furthermore, modern research from the GP literature, graph kernels [@borovitskiy2021matern] and Wasserstein barycentres for GPs [@mallasto2017learning], for example, are supported within `GPJax` but absent from these other packages.
+From both an applied and methodological perspective, GPs are widely employed in the statistics and machine learning communities. High-quality software packages that promote GP modelling are accountable for much of their success. Within the Python community, the three most popular packages for GP modelling are GPFlow [@matthews2017gpflow], GPyTorch [@gardner2018gpytorch], and GPy [@gpy2014]. Despite each of these packages being indispensable tools for the community, none of them support integration with a Jax-based workflow. Both the BayesNewton [@wilkinson2021bayesnewton] and TinyGP [@dfm2021tinygp] packages are built on top of Jax. However, BayesNewton is designed on top of ObJax [@objax2020github], meaning that integration with the wider suite of packages in the Jax ecosystem can be challenging. Meanwhile TinyGP offers excellent integration with inference frameworks such as NumPyro [@phan2019composable] but does not yet support inducing points frameworks (e.g., [@hensman2013gaussian]). `GPJax` exists to resolve this issues. Furthermore, modern research from the GP literature, graph kernels [@borovitskiy2021matern] and Wasserstein barycentres for GPs [@mallasto2017learning], for example, are supported within `GPJax` but absent from these other packages. Finally, the Stheno package [@stheno2022bruinsma] supports a Jax backend along with TensorFlow, PyTorch and Numpy. Whilst this allows GPs to be easily linked with a wider Jax workflow, the benefit of GPJax in contrast is a pure Jax codebase whereas Stheno requires the use of a custom linear algebra framework.
 
-For completeness, packages written for languages other than Python include GPML [@rasmussen2010gaussian] in Matlab, GaussianProcesses.jl [@fairbrother2022gaussianprocesses], AugmentedGaussianProcesses.jl [@fajou20a] and Stheno [@stheno2022tebbutt] all in Julia. Within R, there exists the GPVecchia [@katzfuss2021general] and DiceKriging [@roustant2012dicekriging] packages.
+For completeness, packages written for languages other than Python include GPML [@rasmussen2010gaussian] and GPStuff [@vanhatalo2013gpstuff] in Matlab. An R port also exists for GPStuff. Within Julia, there exists theGaussianProcesses.jl [@fairbrother2022gaussianprocesses], AugmentedGaussianProcesses.jl [@fajou20a] and Stheno.jl [@stheno2022tebbutt] pacakges.
 
[…]
