+++ paper.bib
@@ -7,255 +7,248 @@
 }
 
 @article{matthews2017gpflow,
-  title={{GPflow}: A {Gaussian} Process Library using {TensorFlow}.},
-  author={Matthews, Alexander G de G and Van Der Wilk, Mark and Nickson, Tom and Fujii, Keisuke and Boukouvalas, Alexis and Le{\'o}n-Villagr{\'a}, Pablo and Ghahramani, Zoubin and Hensman, James},
-  journal={Journal of Machine Learning Research},
-  volume={18},
-  number={40},
-  pages={1--6},
-  year={2017}
+  title   = {{GPflow}: A {Gaussian} Process Library using {TensorFlow}.},
+  author  = {Matthews, Alexander G de G and Van Der Wilk, Mark and Nickson, Tom and Fujii, Keisuke and Boukouvalas, Alexis and Le{\'o}n-Villagr{\'a}, Pablo and Ghahramani, Zoubin and Hensman, James},
+  journal = {Journal of Machine Learning Research},
+  volume  = {18},
+  number  = {40},
+  pages   = {1--6},
+  year    = {2017}
 }
 
 @inproceedings{abadi2016tensorflow,
-  title={{TensorFlow}: A System for Large-Scale Machine Learning},
-  author={Abadi, Mart{\'\i}n and Barham, Paul and Chen, Jianmin and Chen, Zhifeng and Davis, Andy and Dean, Jeffrey and Devin, Matthieu and Ghemawat, Sanjay and Irving, Geoffrey and Isard, Michael and others},
-  booktitle={12th {USENIX} symposium on operating systems design and implementation (OSDI 16)},
-  pages={265--283},
-  year={2016}
+  title     = {{TensorFlow}: A System for Large-Scale Machine Learning},
+  author    = {Abadi, Mart{\'\i}n and Barham, Paul and Chen, Jianmin and Chen, Zhifeng and Davis, Andy and Dean, Jeffrey and Devin, Matthieu and Ghemawat, Sanjay and Irving, Geoffrey and Isard, Michael and others},
+  booktitle = {12th {USENIX} symposium on operating systems design and implementation (OSDI 16)},
+  pages     = {265--283},
+  year      = {2016}
 }
 
 
 @article{gardner2018gpytorch,
-  title={{G{p}y{t}orch}: Blackbox matrix-matrix {G}aussian process inference with {GPU} acceleration},
-  author={Gardner, Jacob and Pleiss, Geoff and Weinberger, Kilian Q and Bindel, David and Wilson, Andrew G},
-  journal={Advances in neural information processing systems},
-  volume={31},
-  year={2018}
+  title   = {{G{p}y{t}orch}: Blackbox matrix-matrix {G}aussian process inference with {GPU} acceleration},
+  author  = {Gardner, Jacob and Pleiss, Geoff and Weinberger, Kilian Q and Bindel, David and Wilson, Andrew G},
+  journal = {Advances in neural information processing systems},
+  volume  = {31},
+  year    = {2018}
 }
 
 
 @software{deepmind2020jax,
-  title = {The {D}eep{M}ind {JAX} {E}cosystem},
+  title  = {The {D}eep{M}ind {JAX} {E}cosystem},
[…]
