### 2022-06-20 thomaspinder/GPJax#76 @thomaspinder
Hey @theorashid. I've just update the paper to reflect your most recent comments. My labmate works exclusively in R and tells me all of his GP code is written from scratch. Besides Stan, he was also unaware of any _popular_ R packages for GPs. With that in mind, I've left the R acknowledgements to the GPStuff port that you've flagged above.
