### 2022-06-20 thomaspinder/GPJax#76 @theorashid
Looks like a minor typo "theGaussianProcesses,jl". Worth doing a final check that the references are rendering as you expect too

Once that is fixed, I think we just need to wait for the other reviewer's comments.
