### 2022-06-20 6c26920
Update paper.md
Typos and fluency.
