### 2022-06-20 dc20335
Update paper.md
Consistency.
