### 2022-06-21 thomaspinder/GPJax#76 @Daniel-Dodd
Hey @theorashid. Many thanks for your constructive comments. The NumPyro idea is cool!

@theorashid, @thomaspinder, I have conducted a pass through the paper with fresh eyes and have condensed it down. All changes are on the master branch (apologies - forgot to create a PR - the commits in order are 6c2692034fb685316ef456b1c17f487c990873fa, dc20335dc4dd922a2bbe820eaea95a2512473448 & 11806f61e66a8b3ce0a98894f095d4c2e81a0242). The major change was collapsing the `Summary` section into a single paragraph (to alleviate repetitions) and I addressed the typo you mentioned above @theorashid. Keen to hear both your thoughts on this.

@thomaspinder, please are you able to correct the formatting of `(e.g., (Hensman et al., 2013))` to `(e.g., Hensman et al., 2013)`?    (I am unsure how to do this...)
