+++ paper.bib
@@ -119,7 +119,7 @@
 
 @misc{gpy2014,
   author    = {{GPy}},
-  title     = {{GPy}: A {Gaussian} process framework in python},
+  title     = {{GPy}: A {Gaussian} process framework in {P}ython},
   url       = {http://github.com/SheffieldML/GPy},
   year      = {2012},
   publisher = {Git{H}ub},
