### 2022-06-21 thomaspinder/GPJax#76 @thomaspinder
> Looks like a minor typo "theGaussianProcesses,jl". Worth doing a final check that the references are rendering as you expect too
>
> Once that is fixed, I think we just need to wait for the other reviewer's comments.

@theorashid Nice catch. The typo has been fixed above by Daniel.

@Daniel-Dodd happy with your changes and the references are now updated, as per your request.
