### 2022-06-24 d0fe04b
Fix memory/computation bug in Kxx_diag + tests for TypeError raises
