+++ tests/test_variational_inference.py
@@ -102,3 +102,15 @@ def test_collapsed_vi(n_datapoints, n_inducing_points, jit_fns):
     grads = jax.grad(elbo_fn)(params)
     assert isinstance(grads, tp.Dict)
     assert len(grads) == len(params)
+
+
+
+    # We should raise an error for non-Collapsed variational families:
+    with pytest.raises(TypeError):
+        q = gpx.variational_families.VariationalGaussian(prior=prior, inducing_inputs=inducing_inputs)
+        gpx.variational_inference.AbstractVariationalInference(posterior=post, vartiational_family=q)
+
+        # We should raise an error for non-Gaussian lilkelihoods:
+    with pytest.raises(TypeError):
+        q = gpx.variational_families.CollapsedVariationalGaussian(prior=prior, inducing_inputs=inducing_inputs)
+        gpx.variational_inference.AbstractVariationalInference(posterior=prior * gpx.Bernoulli(D.n), vartiational_family=q)
