### 2022-06-24 0a9cf51
Update test_variational_inference.py
