### 2022-06-24 b4fdd75
Update test_variational_inference.py
