### 2022-06-26 8bf35a8
Minimal working example complete.
