+++ gpjax/natural_gradients.py
@@ -1,23 +1,35 @@
-from copy import deepcopy
-from this import d
 import typing as tp
+from copy import deepcopy
+
+import distrax as dx
 import jax.numpy as jnp
 import jax.scipy as jsp
-import distrax as dx
 from jax import lax, value_and_grad
 
 from .config import get_defaults
-from .variational_families import AbstractVariationalFamily, ExpectationVariationalGaussian, NaturalVariationalGaussian
-from .variational_inference import StochasticVI
-from .utils import I
 from .gps import AbstractPosterior
-from .types import Dataset, Array
-from .parameters import build_identity, transform, build_trainables_false, build_trainables_true, trainable_params
+from .parameters import (
+    build_identity,
+    build_trainables_false,
+    build_trainables_true,
+    trainable_params,
+    transform,
+)
+from .types import Array, Dataset
+from .utils import I
+from .variational_families import (
+    AbstractVariationalFamily,
+    ExpectationVariationalGaussian,
+    NaturalVariationalGaussian,
+)
+from .variational_inference import StochasticVI
 
 DEFAULT_JITTER = get_defaults()["jitter"]
 
 
-def natural_to_expectation(natural_moments: dict, jitter: float = DEFAULT_JITTER) -> dict:
+def natural_to_expectation(
+    natural_moments: dict, jitter: float = DEFAULT_JITTER
+) -> dict:
     """
     Converts natural parameters to expectation parameters.
     Args:
@@ -26,11 +38,11 @@ def natural_to_expectation(natural_moments: dict, jitter: float = DEFAULT_JITTER
     Returns:
         tp.Dict: A dictionary of Gaussian moments under the expectation parameterisation.
     """
-    
+
     natural_matrix = natural_moments["natural_matrix"]
     natural_vector = natural_moments["natural_vector"]
     m = natural_vector.shape[0]
-    
+
     # S⁻¹ = -2θ₂
     S_inv = -2 * natural_matrix
     S_inv += I(m) * jitter
@@ -43,23 +55,27 @@ def natural_to_expectation(natural_moments: dict, jitter: float = DEFAULT_JITTER
 
     # S = CᵀC
     S = jnp.matmul(C.T, C)
-    
+
     # μ = Sθ₁
     mu = jnp.matmul(S, natural_vector)
-    
+
     # η₁ = μ
     expectation_vector = mu
-    
+
     # η₂ = S + μ μᵀ
     expectation_matrix = S + jnp.matmul(mu, mu.T)
-    
-    return {"expectation_vector": expectation_vector, "expectation_matrix": expectation_matrix}
+
+    return {
+        "expectation_vector": expectation_vector,
+        "expectation_matrix": expectation_matrix,
+    }
 
 
-def _expectation_elbo(posterior: AbstractPosterior,
-            variational_family: AbstractVariationalFamily,
-            train_data: Dataset,
-            ) -> tp.Callable[[dict, Dataset], float]:
+def _expectation_elbo(
+    posterior: AbstractPosterior,
+    variational_family: AbstractVariationalFamily,
+    train_data: Dataset,
+) -> tp.Callable[[dict, Dataset], float]:
     """
     Construct evidence lower bound (ELBO) for varational Gaussian under the expectation parameterisation.
     Args:
@@ -68,9 +84,10 @@ def _expectation_elbo(posterior: AbstractPosterior,
     Returns:
         Callable: A function that computes ELBO.
     """
-    evg = ExpectationVariationalGaussian(prior= variational_family.prior, 
-                                        inducing_inputs = variational_family.inducing_inputs,
-                                        )
+    evg = ExpectationVariationalGaussian(
+        prior=variational_family.prior,
[…]
