### 2022-06-27 f2d856a
Add likelihood to collapsed family, update notebook, add reference. Update tests.
