+++ tests/test_variational_inference.py
@@ -1,4 +1,5 @@
 import typing as tp
+
 import jax
 import jax.numpy as jnp
 import jax.random as jr
@@ -6,16 +7,21 @@ import pytest
 
 import gpjax as gpx
 
+
 def test_abstract_variational_inference():
     prior = gpx.Prior(kernel=gpx.RBF())
     lik = gpx.Gaussian(num_datapoints=20)
     post = prior * lik
     n_inducing_points = 10
     inducing_inputs = jnp.linspace(-5.0, 5.0, n_inducing_points).reshape(-1, 1)
-    vartiational_family = gpx.VariationalGaussian(prior=prior, inducing_inputs=inducing_inputs)
+    vartiational_family = gpx.VariationalGaussian(
+        prior=prior, inducing_inputs=inducing_inputs
+    )
 
     with pytest.raises(TypeError):
-        gpx.variational_inference.AbstractVariationalInference(posterior=post, vartiational_family=vartiational_family)
+        gpx.variational_inference.AbstractVariationalInference(
+            posterior=post, vartiational_family=vartiational_family
+        )
 
 
 def get_data_and_gp(n_datapoints):
@@ -39,12 +45,12 @@ def test_stochastic_vi(n_datapoints, n_inducing_points, n_test, whiten, diag, ji
     inducing_inputs = jnp.linspace(-5.0, 5.0, n_inducing_points).reshape(-1, 1)
 
     if whiten is True:
-        q = gpx.WhitenedVariationalGaussian(prior = prior,
-        inducing_inputs=inducing_inputs, diag=diag
+        q = gpx.WhitenedVariationalGaussian(
+            prior=prior, inducing_inputs=inducing_inputs, diag=diag
         )
     else:
-        q = gpx.VariationalGaussian(prior = prior,
-        inducing_inputs=inducing_inputs, diag=diag
+        q = gpx.VariationalGaussian(
+            prior=prior, inducing_inputs=inducing_inputs, diag=diag
         )
 
     svgp = gpx.StochasticVI(posterior=post, variational_family=q)
@@ -71,13 +77,18 @@ def test_stochastic_vi(n_datapoints, n_inducing_points, n_test, whiten, diag, ji
     assert isinstance(grads, tp.Dict)
     assert len(grads) == len(params)
 
+
 @pytest.mark.parametrize("n_datapoints, n_inducing_points", [(10, 2), (100, 10)])
 @pytest.mark.parametrize("jit_fns", [False, True])
 def test_collapsed_vi(n_datapoints, n_inducing_points, jit_fns):
     D, post, prior = get_data_and_gp(n_datapoints)
+
+    likelihood = gpx.Gaussian(num_datapoints=n_datapoints)
     inducing_inputs = jnp.linspace(-5.0, 5.0, n_inducing_points).reshape(-1, 1)
 
-    q = gpx.variational_families.CollapsedVariationalGaussian(prior=prior, inducing_inputs=inducing_inputs)
+    q = gpx.variational_families.CollapsedVariationalGaussian(
+        prior=prior, likelihood=likelihood, inducing_inputs=inducing_inputs
+    )
 
     sgpr = gpx.variational_inference.CollapsedVI(posterior=post, variational_family=q)
     assert sgpr.posterior.prior == post.prior
@@ -103,14 +114,18 @@ def test_collapsed_vi(n_datapoints, n_inducing_points, jit_fns):
     assert isinstance(grads, tp.Dict)
     assert len(grads) == len(params)
 
-
-
     # We should raise an error for non-Collapsed variational families:
     with pytest.raises(TypeError):
-        q = gpx.variational_families.VariationalGaussian(prior=prior, inducing_inputs=inducing_inputs)
+        q = gpx.variational_families.VariationalGaussian(
[…]
