+++ docs/refs.bib
@@ -67,3 +67,15 @@
   year={2016},
   organization={PMLR}
 }
+
+
+@InProceedings{titsias2009,
+  title = 	 {Variational Learning of Inducing Variables in Sparse Gaussian Processes},
+  author = 	 {Titsias, Michalis},
+  booktitle = 	 {Proceedings of the Twelth International Conference on Artificial Intelligence and Statistics},
+  pages = 	 {567--574},
+  year = 	 {2009},
+  volume = 	 {5},
+  series = 	 {Proceedings of Machine Learning Research},
+  publisher =    {PMLR},
+}
