### 2022-06-27 c2332d0
Update gotchas.pct.py
