+++ docs/nbs/gotchas.pct.py
@@ -18,44 +18,44 @@
 # %% [markdown]
 # # Common Gotchas in GPJax
 #
-# GPJax is an incredibly flexible framework due to it's low-level of abstraction. However, by providing a large degree of flexibility to users, it is possbile for pain points to be encountered when developing custom models. In this notebook we try to shed some light on these _gothcas_, providing explanations for their presence and giving appropriate solutions where applicable.
+# GPJax is incredibly flexible due to it's low-level of abstraction. However, by providing a significant degree of flexibility to users, pain points can be encountered when developing custom models. In this notebook, we try to shed some light on these, _gothcas_, providing explanations for their presence and giving appropriate solutions where applicable.
 #
-# We will not attempt to cover all Jax based edge cases in this notebook as they are covered superbly in the [Jax - The Sharp Bits](https://jax.readthedocs.io/en/latest/notebooks/Common_Gotchas_in_JAX.html) notebook.
+# We will not attempt to cover all Jax-based edge cases in this notebook as they are covered superbly in the [Jax - The Sharp Bits](https://jax.readthedocs.io/en/latest/notebooks/Common_Gotchas_in_JAX.html) notebook.
 #
-# _This list is non-exhaustive. If you encounter your own confusion, then please consider opening [contributing](https://github.com/thomaspinder/GPJax/blob/master/CONTRIBUTING.md) a PR to this notebook._
+# _This list is non-exhaustive. If you encounter confusion, please consider opening [contributing](https://github.com/thomaspinder/GPJax/blob/master/CONTRIBUTING.md) a PR to this notebook._
 
 # %% [markdown]
 # ## Chex Dataclasses
 #
-# Dataclasses are supported in base Python since version 3.7. However, for a dataclass object to work with Jax, they must be a registered PyTree. This is where [Chex](https://github.com/deepmind/chex) comes in. Chex is Jax library that provides an implementation of Python's `dataclass` object that is a valid PyTree. Using dataclasses inplace of regular Python classes removes a signifacnt amount of boilerplate when defining new and thus keeps the underlying codebase more concise. However, there are a few specifics that one should be mindfule of when using dataclasses.
+# Dataclasses have been supported in base Python since version 3.7. However, for dataclass objects to work with Jax, they must be a registered PyTree. This is where [Chex](https://github.com/deepmind/chex) comes in. Chex is a Jax library, that provides an implementation of Python's `dataclass` object that is a valid PyTree. Using dataclasses in place of regular Python classes removes a significant amount of boilerplate when defining new and thus keeps the underlying codebase more concise. However, there are a few specifics that one should be mindful of when using dataclasses.
 #
 # ### Instantiation
 #
-# When instantiating a dataclass, a requirement is that all arguments are keyword arguments. As an example, consider the instantiation of a GP prior that is parameterised by an RBF kernel:
+# When instantiating a dataclass, a requirement is that all arguments are keyword arguments. As an example, consider the instantiation of a GP prior, parameterised by an RBF kernel:
 # ```diff
 # ❌ prior = gpx.Prior(gpx.RBF())
 # ✅ prior = gpx.Prior(kernel=gpx.RBF())
 # ```
-# Running the former will error as the kernel argument has been supplied without an associated keyword name. Takeaway: __when instantiating GPJax object, all arguments must be named arguments__.
+# Running the former will result in an error because the kernel argument is missing its associated keyword. Takeaway: __when instantiating a GPJax object, all arguments must be named arguments__.
 #
 # ### Inheritance
 #
-# Dataclasses are just regular classes under the hood. However, there is some nuance when inheriting dataclasses from one another. Namely this boils down to the inheritance order being in reverse.
[…]
