### 2022-06-30 782297e
Update mean_functions.py
Name for the constant in the constant mean object, "variance", does not make sense. Have renamed this to "constant". Have also reshaped the default value to be a 1 dimensional array for consistency with specifying kernel hyperparameters in GPJax.
