### 2022-06-30 cec132c
Captialise Gaussian
