+++ docs/refs.bib
@@ -70,7 +70,7 @@
 
 
 @InProceedings{titsias2009,
-  title = 	 {Variational Learning of Inducing Variables in Sparse Gaussian Processes},
+  title = 	 {Variational Learning of Inducing Variables in Sparse {G}aussian Processes},
   author = 	 {Titsias, Michalis},
   booktitle = 	 {Proceedings of the Twelth International Conference on Artificial Intelligence and Statistics},
   pages = 	 {567--574},
