### 2022-06-30 #4455 @gpleiss
- [x] Repository: Is the source code for this software available at the [repository url](https://github.com/thomaspinder/GPJax)?
- [x] License: Does the repository contain a plain-text LICENSE file with the contents of an [OSI approved](https://opensource.org/licenses/alphabetical) software license?
- [x] Version: Does the release version given match the GitHub release (v0.4.6)?
- [x] Authorship: Has the submitting author (@thomaspinder) made major contributions to the software? Does the full list of paper authors seem appropriate and complete?

  - [ ] **Issue**: GPU setup is a bit challenging. To get GPU support, I have to `pip install --update jax[cuda]` before installing GPJax. However, GPJax's requirements lock the Jax version to a specific release (0.3.2). I got installation to work, but it's a bit clunky. I would recommend that the authors:
     - [ ] Create a CUDA specific install (e.g. `pip install "gpjax[cuda]"`) that would install `jax[cuda])
     - [ ] Make the requirements less strict (unless there is a good reason to do so?
  - It's possible that there's an easier way to get CUDA support (I haven't used JAX much!). @thomaspinder I'm curious to hear your thoughts.
  - [ ] **Nice to have**: Since the software is designed to pair well with the GPML book, it would be beneficial to include more of the kernels discussed in that book. Right now the set of kernels is limited (although the included kernels are arguably the most useful). Off the top of my head, it would be useful to include the RQ kernel, the periodic kernel, the linear kernel, and maybe something more advanced like the Gibbs kernel.
     - The authors include good documentation for writing custom kernels, and it looks like it would be easy for a user to implement any of these kernels themselves. However, I think the library could benefit from including some of these more common kernels to reduce user overhead (especially for new users).

  - [ ] **Nice to have**: It would be good to have at least one code example with higher dimensional data (e.g. a UCI dataset). I recognize that 1D data is much nicer to visualize, but having a "real dataset" would make for a more comprehensive example, especially for very new users.
  - Very thorough!
- [x] Automated tests: Are there automated tests or manual steps described so that the function of the software can be verified?
  - [ ] **Issue**: You should set up a GitHub action to automatically run your tests for PRs and commits to master.
  - [ ] **Suggestion**: It could be useful to have some larger end-to-end tests (beyond the unit tests) to make sure that all the pieces stay glued together. The pyro folks have a very clever way of doing this: [they run all of their example notebooks as a smoke test](https://github.com/pyro-ppl/pyro/actions/runs/2565968752/workflow#L97).

- [x] Summary: Has a clear description of the high-level functionality and purpose of the software for a diverse, non-specialist audience been provided?
- [x] A statement of need: Does the paper have a section titled 'Statement of Need' that clearly states what problems the software is designed to solve and who the target audience is?
  - [ ] **Nit:** the "statement of need" in the paper doesn't actually address the need; that comes in the "wider software ecosystem" section. I would suggest that the authors reorganize these two sections?
- [x] State of the field: Do the authors describe how this software compares to other commonly-used packages?
- [x] Quality of writing: Is the paper well written (i.e., it does not require editing for structure, language, or writing quality)?
  - I didn't read too closely for typos
- [x] References: Is the list of references complete, and is everything cited appropriately that should be cited (e.g., papers, datasets, software)? Do references in the text use the proper [citation syntax](https://pandoc.org/MANUAL.html#extension-citations)?

I'm really excited about this library, excellent work! Beyond my (minor) comments above, this package will be a great resource for the GP community, and it has been built to high open source standards.
