### 2022-06-30 #4455 @gpleiss
@dfm @thomaspinder I'll post the actionable items from my review in @theorashid 's [review issue](https://github.com/thomaspinder/GPJax/issues/76)
