### 2022-06-30 thomaspinder/GPJax#76 @gpleiss
Hi @thomaspinder and @Daniel-Dodd, here's some additional items from my review:

- [ ] **Issue**: GPU setup is a bit challenging. To get GPU support, I have to `pip install --update jax[cuda]` before installing GPJax. However, GPJax's requirements lock the Jax version to a specific release (0.3.2). I got installation to work, but it's a bit clunky. I would recommend that the authors:
  - [ ] Create a CUDA specific install (e.g. `pip install "gpjax[cuda]"`) that would install `jax[cuda])
  - [ ] Make the requirements less strict (unless there is a good reason to do so?
  - It's possible that there's an easier way to get CUDA support (I haven't used JAX much!). @thomaspinder I'm curious to hear your thoughts.
- [ ] **Nice to have**: Since the software is designed to pair well with the GPML book, it would be beneficial to include more of the kernels discussed in that book. Right now the set of kernels is limited (although the included kernels are arguably the most useful). Off the top of my head, it would be useful to include the RQ kernel, the periodic kernel, the linear kernel, and maybe something more advanced like the Gibbs kernel.
- [ ] **Nice to have**: It would be good to have at least one code example with higher dimensional data (e.g. a UCI dataset). I recognize that 1D data is much nicer to visualize, but having a "real dataset" would make for a more comprehensive example, especially for very new users.
- [ ] **Issue**: You should set up a GitHub action to automatically run your tests for PRs and commits to master.
- [ ] **Suggestion**: It could be useful to have some larger end-to-end tests (beyond the unit tests) to make sure that all the pieces stay glued together. The pyro folks have a very clever way of doing this: [they run all of their example notebooks as a smoke test](https://github.com/pyro-ppl/pyro/actions/runs/2565968752/workflow#L97).
- [ ] **Nit:** the "statement of need" in the paper doesn't actually address the need; that comes in the "wider software ecosystem" section. I would suggest that the authors reorganize these two sections?

Overall, I'm really excited about this library, excellent work! I think the things marked as **issue** are important, but everything else is merely suggestion.
