### 2022-07-01 thomaspinder/GPJax#76 @thomaspinder
Thanks @gpleiss for such helpful comments! I’m currently away at ISBA, so I‘ll address your comments on Monday once I’m home. Having read your comments though, I’m confident that we can resolve them all promptly.
