### 2022-07-01 01918aa
Add more kernels.
Need to:

- Finish periodic kernel.
- Incorporate full tests for the others.
