+++ gpjax/kernels.py
@@ -170,6 +170,72 @@ class RBF(Kernel):
         K = params["variance"] * jnp.exp(-0.5 * squared_distance(x, y))
         return K.squeeze()
 
+@dataclass(repr=False)
+class PoweredExponential(Kernel):
+    """The powered exponential family of kernels.
+
+    Key reference is Diggle and Ribeiro (2007) - "Model-based Geostatistics".
+    
+    """
+
+    name: Optional[str] = "Powered exponential"
+
+    def __post_init__(self):
+        self.ndims = 1 if not self.active_dims else len(self.active_dims)
+        self._params = {
+            "lengthscale": jnp.array([1.0] * self.ndims),
+            "variance": jnp.array([1.0]),
+            "power": jnp.array([1.0]),
+        }
+
+    def __call__(self, x: jnp.DeviceArray, y: jnp.DeviceArray, params: dict) -> Array:
+        """Evaluate the kernel on a pair of inputs :math:`(x, y)` with length-scale parameter :math:`\ell`, :math:`\sigma` and power :math:`\kappa`.
+
+        .. math::
+            k(x, y) = \\sigma^2 \\exp \\Bigg( - \\Big( \\frac{\\lVert x - y \\rVert^2}{\\ell^2} \\Big)^\\kappa \\Bigg)
+
+        Args:
+            x (jnp.DeviceArray): The left hand argument of the kernel function's call.
+            y (jnp.DeviceArray): The right hand argument of the kernel function's call
+            params (dict): Parameter set for which the kernel should be evaluated on.
+
+        Returns:
+            Array: The value of :math:`k(x, y)`
+        """
+        x = self.slice_input(x) / params["lengthscale"]
+        y = self.slice_input(y) / params["lengthscale"]
+        K = params["variance"] * jnp.exp(-euclidean_distance(x, y) ** params["power"])
+        return K.squeeze()
+
+
+@dataclass(repr=False)
+class Linear(Kernel):
+    """The linear kernel."""
+
+    name: Optional[str] = "Linear"
+
+    def __post_init__(self):
+        self.ndims = 1 if not self.active_dims else len(self.active_dims)
+        self._params = {"variance": jnp.array([1.0]), "bias": jnp.array([0.0])}
+
+    def __call__(self, x: jnp.DeviceArray, y: jnp.DeviceArray, params: dict) -> Array:
+        """Evaluate the kernel on a pair of inputs :math:`(x, y)` with bias parameter :math:`b` and variance :math:`\sigma`
+
+        .. math::
+            k(x, y) = \\sigma^2 x^{T}y + b
+
+        Args:
+            x (jnp.DeviceArray): The left hand argument of the kernel function's call.
+            y (jnp.DeviceArray): The right hand argument of the kernel function's call
+            params (dict): Parameter set for which the kernel should be evaluated on.
+        Returns:
+            Array: The value of :math:`k(x, y)`
+        """
+        x = self.slice_input(x)
+        y = self.slice_input(y)
+        K = params["variance"] * jnp.matmul(x.T, y) + params["bias"]
+        return K.squeeze()
+
 
 @dataclass(repr=False)
 class Matern12(Kernel):
@@ -188,7 +254,7 @@ class Matern12(Kernel):
         """Evaluate the kernel on a pair of inputs :math:`(x, y)` with length-scale parameter :math:`\ell` and variance :math:`\sigma`
 
         .. math::
-            k(x, y) = \\sigma^2 \\exp \\Bigg( -\\frac{\\lvert x-y \\rvert}{\\ell}  \\Bigg)
[…]
