### 2022-07-04 430652b
Additional VI tests
