+++ tests/test_variational_inference.py
@@ -1,4 +1,5 @@
 import typing as tp
+
 import jax
 import jax.numpy as jnp
 import jax.random as jr
@@ -6,21 +7,27 @@ import pytest
 
 import gpjax as gpx
 
+
 def test_abstract_variational_inference():
     prior = gpx.Prior(kernel=gpx.RBF())
     lik = gpx.Gaussian(num_datapoints=20)
     post = prior * lik
     n_inducing_points = 10
     inducing_inputs = jnp.linspace(-5.0, 5.0, n_inducing_points).reshape(-1, 1)
-    vartiational_family = gpx.VariationalGaussian(prior=prior, inducing_inputs=inducing_inputs)
+    vartiational_family = gpx.VariationalGaussian(
+        prior=prior, inducing_inputs=inducing_inputs
+    )
 
     with pytest.raises(TypeError):
-        gpx.variational_inference.AbstractVariationalInference(posterior=post, vartiational_family=vartiational_family)
+        gpx.variational_inference.AbstractVariationalInference(
+            posterior=post, vartiational_family=vartiational_family
+        )
 
 
-def get_data_and_gp(n_datapoints):
+def get_data_and_gp(n_datapoints, point_dim):
     x = jnp.linspace(-5.0, 5.0, n_datapoints).reshape(-1, 1)
     y = jnp.sin(x) + jr.normal(key=jr.PRNGKey(123), shape=x.shape) * 0.1
+    x = jnp.hstack([x] * point_dim)
     D = gpx.Dataset(X=x, y=y)
 
     p = gpx.Prior(kernel=gpx.RBF())
@@ -34,17 +41,21 @@ def get_data_and_gp(n_datapoints):
 @pytest.mark.parametrize("whiten", [True, False])
 @pytest.mark.parametrize("diag", [True, False])
 @pytest.mark.parametrize("jit_fns", [False, True])
-def test_stochastic_vi(n_datapoints, n_inducing_points, n_test, whiten, diag, jit_fns):
-    D, post, prior = get_data_and_gp(n_datapoints)
+@pytest.mark.parametrize("point_dim", [1, 2])
+def test_stochastic_vi(
+    n_datapoints, n_inducing_points, n_test, whiten, diag, jit_fns, point_dim
+):
+    D, post, prior = get_data_and_gp(n_datapoints, point_dim)
     inducing_inputs = jnp.linspace(-5.0, 5.0, n_inducing_points).reshape(-1, 1)
+    inducing_inputs = jnp.hstack([inducing_inputs] * point_dim)
 
     if whiten is True:
-        q = gpx.WhitenedVariationalGaussian(prior = prior,
-        inducing_inputs=inducing_inputs, diag=diag
+        q = gpx.WhitenedVariationalGaussian(
+            prior=prior, inducing_inputs=inducing_inputs, diag=diag
         )
     else:
-        q = gpx.VariationalGaussian(prior = prior,
-        inducing_inputs=inducing_inputs, diag=diag
+        q = gpx.VariationalGaussian(
+            prior=prior, inducing_inputs=inducing_inputs, diag=diag
         )
 
     svgp = gpx.StochasticVI(posterior=post, variational_family=q)
@@ -71,13 +82,18 @@ def test_stochastic_vi(n_datapoints, n_inducing_points, n_test, whiten, diag, ji
     assert isinstance(grads, tp.Dict)
     assert len(grads) == len(params)
 
+
 @pytest.mark.parametrize("n_datapoints, n_inducing_points", [(10, 2), (100, 10)])
 @pytest.mark.parametrize("jit_fns", [False, True])
-def test_collapsed_vi(n_datapoints, n_inducing_points, jit_fns):
[…]
