+++ requirements.txt
@@ -1,5 +1,5 @@
-jax==0.3.2
-jaxlib==0.3.2
+jax>=0.3.2
+jaxlib>=0.3.2
 optax>=0.1.0
 chex==0.1.3
 distrax>=0.1.2
