+++ docs/index.rst
@@ -39,8 +39,8 @@ For intuition of GPJax's exposed API, a simple example that derives the Gaussian
 
     nbs/regression
     nbs/classification
-    nbs/sparse_regression
-    nbs/sgpr
+    nbs/uncollapsed_vi
+    nbs/collapsed_vi
     nbs/graph_kernels
     nbs/barycentres
     nbs/haiku
