+++ .github/workflows/publish.yml
@@ -1,0 +1,49 @@
+name: Publish Python 🐍 distributions 📦 to PyPI and TestPyPI
+
+# publish wheels and source to test.pypi.org on releases/* branch
+# publish wheels and source to pypi.org when pushing v* tag on master or releasees/* branch
+
+on:
+  push:
+    branches:
+      - master
+      - releases/*
+    tags:
+      - v*
+
+jobs:
+  build-and-publish:
+    runs-on: ubuntu-latest
+    steps:
+      - name: Checkout
+      - uses: actions/checkout@v2
+        with:
+          fetch-depth: 20
+      - name: Fetch release tag
+        run: git fetch --depth=1 origin +refs/tags/*:refs/tags/*
+      - name: Setup Python 3.8
+      - uses: actions/setup-python@v2
+        name: Install Python
+        with:
+          python-version: 3.8
+      - name: Build sdist and wheel
+        run: |
+          python -m pip install -U pip
+          python -m pip install -U build
+          python -m build .
+      - name: Store
+      - uses: actions/upload-artifact@v2
+        with:
+          path: dist/*
+      - name: Publish to PyPi
+        if: github.event_name == 'push' && startsWith(github.ref, 'refs/tags')
+        steps:
+          - uses: actions/download-artifact@v2
+          with:
+            name: artifact
+            path: dist
+
+            - uses: pypa/gh-action-pypi-publish@v1.5.0
+              with:
+                user: __token__
+                password: ${{ secrets.PYPI_API_TOKEN }}
