+++ .github/workflows/workflow-master.yml
@@ -31,37 +31,3 @@ jobs:
           token: ${{ secrets.CODECOV_TOKEN }}
           file: ./coverage.xml
           flags: unittests
-
-  build:
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@v2
-        with:
-          fetch-depth: 0
-      - uses: actions/setup-python@v2
-        name: Install Python
-        with:
-          python-version: "3.8"
-      - name: Build sdist and wheel
-        run: |
-          python -m pip install -U pip
-          python -m pip install -U build
-          python -m build .
-      - uses: actions/upload-artifact@v2
-        with:
-          path: dist/*
-
-  upload_pypi:
-    needs: [codecov, build]
-    runs-on: ubuntu-latest
-    if: github.event_name == 'push' && startsWith(github.ref, 'refs/tags')
-    steps:
-      - uses: actions/download-artifact@v2
-        with:
-          name: artifact
-          path: dist
-
-      - uses: pypa/gh-action-pypi-publish@v1.5.0
-        with:
-          user: __token__
-          password: ${{ secrets.PYPI_API_TOKEN }}
