+++ .github/workflows/publish.yml
@@ -1,49 +1,47 @@
-name: Publish Python 🐍 distributions 📦 to PyPI and TestPyPI
-
-# publish wheels and source to test.pypi.org on releases/* branch
-# publish wheels and source to pypi.org when pushing v* tag on master or releasees/* branch
+#follows https://packaging.python.org/guides/publishing-package-distribution-releases-using-github-actions-ci-cd-workflows/
+name: Publish Python 🐍  distributions 📦 to PyPI and TestPyPI
 
 on:
-  push:
-    branches:
-      - master
-      - releases/*
-    tags:
-      - v*
+  workflow_dispatch:
+    inputs:
+      releasetype:
+        description: "Is this a 'release' or a 'test'?"
+        required: true
+        default: "test"
 
 jobs:
-  build-and-publish:
+  build-n-publish:
+    name: Build and publish Python 🐍 distributions 📦 to PyPI and TestPyPI
     runs-on: ubuntu-latest
+
     steps:
-      - name: Checkout
-      - uses: actions/checkout@v2
-        with:
-          fetch-depth: 20
-      - name: Fetch release tag
-        run: git fetch --depth=1 origin +refs/tags/*:refs/tags/*
-      - name: Setup Python 3.8
-      - uses: actions/setup-python@v2
-        name: Install Python
+      - name: Display phase
+        run: echo "Deploying "${{github.event.inputs.releasetype}}" $INPUT_RELEASETYPE to PyPI"
+
+      - uses: actions/checkout@master
+
+      - name: Set up Python 3.8
+        uses: actions/setup-python@v1
         with:
           python-version: 3.8
-      - name: Build sdist and wheel
-        run: |
-          python -m pip install -U pip
-          python -m pip install -U build
-          python -m build .
-      - name: Store
-      - uses: actions/upload-artifact@v2
-        with:
-          path: dist/*
-      - name: Publish to PyPi
-        if: github.event_name == 'push' && startsWith(github.ref, 'refs/tags')
-        steps:
-          - uses: actions/download-artifact@v2
-          with:
-            name: artifact
-            path: dist
 
-            - uses: pypa/gh-action-pypi-publish@v1.5.0
-              with:
-                user: __token__
-                password: ${{ secrets.PYPI_API_TOKEN }}
+      - name: Build a test source tarball
+        if: github.event.inputs.releasetype == 'test'
+        run: PYTERRIER_VERSION_SUFFIX=".`date +%s`" python setup.py sdist
+
+      - name: Build a release source tarball
+        if: github.event.inputs.releasetype == 'release'
+        run: python setup.py sdist
+
+      - name: Publish distribution 📦 to Test PyPI
+        if: github.event.inputs.releasetype == 'test'
+        uses: pypa/gh-action-pypi-publish@master
+        with:
+          password: ${{ secrets.TEST_PYPI_API_TOKEN }}
+          repository_url: https://test.pypi.org/legacy/
+
+      - name: Publish distribution 📦 to PyPI
+        if: github.event.inputs.releasetype == 'release'
+        uses: pypa/gh-action-pypi-publish@master
+        with:
+          password: ${{ secrets.PYPI_API_TOKEN }}
