### 2022-07-06 0ddddf0
Fix hello world. Make name strings consistent.
- Fix (so that it runs without error) and tidy hello world example.
- Add name strings to whitened  variational family.
- Make name strings follow consistent capitalisation and formatting.
