### 2022-07-06 4fda7e1
Hide index and add GP graphic.
