+++ docs/index.rst
@@ -1,14 +1,18 @@
 :hide-toc: true
 
 
-Welcome to GPJax's documentation!
+Welcome to GPJax!
 =================================
 
-GPJax is a didactic Gaussian process library that supports GPU acceleration and just-in-time compilation. We seek to provide a flexible API as close as possible to how the underlying mathematics is written on paper to enable researchers rapidly prototype and develop new ideas.
+GPJax is a didactic Gaussian process library that supports GPU acceleration and just-in-time compilation. We seek to provide a flexible API as close as possible to how the underlying mathematics is written on paper to enable researchers to rapidly prototype and develop new ideas.
+
+.. image:: ./GP.pdf
+  :width: 800
+  :alt: Alternative text
 
 You can view the source code for GPJax `here on Github <https://github.com/thomaspinder/GPJax>`_.
 
-Gaussian process 'Hello World' example
+'Hello World' example
 -----------------------------------------------------
 
 Defining a Gaussian process posterior is simple as typing the maths we would write on paper.
@@ -26,10 +30,10 @@ Defining a Gaussian process posterior is simple as typing the maths we would wri
 
 To learn more, checkout the `regression notebook <https://gpjax.readthedocs.io/en/latest/nbs/regression.html>`_.
 
-
 .. toctree::
     :maxdepth: 1
     :caption: Getting Started
+    :hidden:
 
     installation
     design
@@ -37,6 +41,7 @@ To learn more, checkout the `regression notebook <https://gpjax.readthedocs.io/e
 .. toctree::
     :maxdepth: 1
     :caption: Examples
+    :hidden:
 
     nbs/regression
     nbs/classification
@@ -50,19 +55,19 @@ To learn more, checkout the `regression notebook <https://gpjax.readthedocs.io/e
 .. toctree::
     :maxdepth: 1
     :caption: Guides
+    :hidden:
 
     nbs/kernels
 
 .. toctree::
     :maxdepth: 1
     :caption: Package Reference
+    :hidden:
 
     api
 
-
 Indices and tables
--------------------
-
+-----------------------------
 * :ref:`genindex`
 * :ref:`modindex`
 * :ref:`search`
