+++ gpjax/variational_families.py
@@ -6,12 +6,13 @@ import jax.numpy as jnp
 import jax.scipy as jsp
 import tensorflow_probability.substrates.jax.bijectors as tfb
 from chex import dataclass
+from jaxtyping import f64
 
 from .config import Identity, Softplus, add_parameter, get_defaults
 from .gps import Prior
 from .kernels import cross_covariance, gram
 from .likelihoods import AbstractLikelihood, Gaussian
-from .types import Array, Dataset
+from .types import Dataset
 from .utils import I, concat_dictionaries
 
 DEFAULT_JITTER = get_defaults()["jitter"]
@@ -54,10 +55,10 @@ class VariationalGaussian(AbstractVariationalFamily):
     """
 
     prior: Prior
-    inducing_inputs: Array
+    inducing_inputs: f64["M D"]
     name: str = "Variational Gaussian"
-    variational_mean: Optional[Array] = None
-    variational_root_covariance: Optional[Array] = None
+    variational_mean: Optional[f64["M Q"]] = None
+    variational_root_covariance: Optional[f64["M M"]] = None
     diag: Optional[bool] = False
     jitter: Optional[float] = DEFAULT_JITTER
 
@@ -93,7 +94,7 @@ class VariationalGaussian(AbstractVariationalFamily):
             },
         )
 
-    def prior_kl(self, params: Dict) -> Array:
+    def prior_kl(self, params: Dict) -> f64["1"]:
         """Compute the KL-divergence between our variational approximation and the Gaussian process prior.
 
         For this variational family, we have KL[q(f(·))||p(·)] = KL[q(u)||p(u)] = KL[ N(μ, S) || N(μz, Kzz) ],
@@ -119,7 +120,7 @@ class VariationalGaussian(AbstractVariationalFamily):
 
         return qu.kl_divergence(pu)
 
-    def predict(self, params: dict) -> Callable[[Array], dx.Distribution]:
+    def predict(self, params: dict) -> Callable[[f64["N D"]], dx.Distribution]:
         """Compute the predictive distribution of the GP at the test inputs t.
 
         This is the integral q(f(t)) = ∫ p(f(t)|u) q(u) du, which can be computed in closed form as:
@@ -142,7 +143,7 @@ class VariationalGaussian(AbstractVariationalFamily):
         Lz = jnp.linalg.cholesky(Kzz)
         μz = self.prior.mean_function(z, params["mean_function"])
 
-        def predict_fn(test_inputs: Array) -> dx.Distribution:
+        def predict_fn(test_inputs: f64["N D"]) -> dx.Distribution:
             t = test_inputs
             n_test = t.shape[0]
             Ktt = gram(self.prior.kernel, t, params["kernel"])
@@ -187,7 +188,7 @@ class WhitenedVariationalGaussian(VariationalGaussian):
 
     name: str = "Whitened variational Gaussian"
 
-    def prior_kl(self, params: Dict) -> Array:
+    def prior_kl(self, params: Dict) -> f64["1"]:
         """Compute the KL-divergence between our variational approximation and the Gaussian process prior.
 
         For this variational family, we have KL[q(f(·))||p(·)] = KL[q(u)||p(u)] = KL[N(μ, S)||N(0, I)].
@@ -207,7 +208,7 @@ class WhitenedVariationalGaussian(VariationalGaussian):
 
         return qu.kl_divergence(pu)
 
-    def predict(self, params: dict) -> Callable[[Array], dx.Distribution]:
+    def predict(self, params: dict) -> Callable[[f64["N D"]], dx.Distribution]:
         """Compute the predictive distribution of the GP at the test inputs t.
 
[…]
