+++ gpjax/kernels.py
@@ -4,8 +4,7 @@ from typing import Callable, Dict, List, Optional, Sequence
 import jax.numpy as jnp
 from chex import dataclass
 from jax import vmap
-
-from gpjax.types import Array
+from jaxtyping import f64
 
 
 ##########################################
@@ -25,7 +24,7 @@ class Kernel:
         self.ndims = 1 if not self.active_dims else len(self.active_dims)
 
     @abc.abstractmethod
-    def __call__(self, x: Array, y: Array, params: dict) -> Array:
+    def __call__(self, x: f64["1 D"], y: f64["1 D"], params: dict) -> f64["1"]:
         """Evaluate the kernel on a pair of inputs.
         Args:
             x (jnp.DeviceArray): The left hand argument of the kernel function's call.
@@ -36,7 +35,7 @@ class Kernel:
         """
         raise NotImplementedError
 
-    def slice_input(self, x: Array) -> Array:
+    def slice_input(self, x: f64["N D"]) -> f64["N Q"]:
         """Select the relevant columns of the supplied matrix to be used within the kernel's evaluation.
         Args:
             x (Array): The matrix or vector that is to be sliced.
@@ -113,7 +112,7 @@ class CombinationKernel(Kernel, _KernelSet):
         """A template dictionary of the kernel's parameter set."""
         return [kernel.params for kernel in self.kernel_set]
 
-    def __call__(self, x: Array, y: Array, params: dict) -> Array:
+    def __call__(self, x: f64["1 D"], y: f64["1 D"], params: dict) -> f64["1"]:
         return self.combination_fn(
             jnp.stack([k(x, y, p) for k, p in zip(self.kernel_set, params)])
         )
@@ -151,7 +150,7 @@ class RBF(Kernel):
             "variance": jnp.array([1.0]),
         }
 
-    def __call__(self, x: jnp.DeviceArray, y: jnp.DeviceArray, params: dict) -> Array:
+    def __call__(self, x: f64["1 D"], y: f64["1 D"], params: dict) -> f64["1"]:
         """Evaluate the kernel on a pair of inputs :math:`(x, y)` with length-scale parameter :math:`\ell` and variance :math:`\sigma`
 
         .. math::
@@ -184,7 +183,7 @@ class Matern12(Kernel):
             "variance": jnp.array([1.0]),
         }
 
-    def __call__(self, x: jnp.DeviceArray, y: jnp.DeviceArray, params: dict) -> Array:
+    def __call__(self, x: f64["1 D"], y: f64["1 D"], params: dict) -> f64["1"]:
         """Evaluate the kernel on a pair of inputs :math:`(x, y)` with length-scale parameter :math:`\ell` and variance :math:`\sigma`
 
         .. math::
@@ -216,7 +215,7 @@ class Matern32(Kernel):
             "variance": jnp.array([1.0]),
         }
 
-    def __call__(self, x: jnp.DeviceArray, y: jnp.DeviceArray, params: dict) -> Array:
+    def __call__(self, x: f64["1 D"], y: f64["1 D"], params: dict) -> f64["1"]:
         """Evaluate the kernel on a pair of inputs :math:`(x, y)` with lengthscale parameter :math:`\ell` and variance :math:`\sigma`
 
         .. math::
@@ -254,7 +253,7 @@ class Matern52(Kernel):
             "variance": jnp.array([1.0]),
         }
 
-    def __call__(self, x: jnp.DeviceArray, y: jnp.DeviceArray, params: dict) -> Array:
[…]
