+++ gpjax/quadrature.py
@@ -2,8 +2,7 @@ from typing import Callable
 
 import jax.numpy as jnp
 import numpy as np
-
-from .types import Array
+from jaxtyping import f64
 
 """The number of Gauss-Hermite points to use for quadrature"""
 DEFAULT_NUM_GAUSS_HERMITE_POINTS = 20
@@ -11,12 +10,12 @@ DEFAULT_NUM_GAUSS_HERMITE_POINTS = 20
 
 def gauss_hermite_quadrature(
     fun: Callable,
-    mean: Array,
-    var: Array,
+    mean: f64["N D"],
+    var: f64["N D"],
     deg: int = DEFAULT_NUM_GAUSS_HERMITE_POINTS,
     *args,
     **kwargs
-) -> Array:
+) -> f64["D"]:
     """Compute Gaussian-Hermite quadrature for a given function. The quadrature points are adjusted through the supplied mean and variance arrays.
 
     Args:
