+++ docs/requirements.txt
@@ -16,5 +16,7 @@ blackjax>=0.8.2
 jaxopt
 dm-haiku
 ipywidgets
+pandas
+scikit-learn
 # Install GPJax istself
 .
