### 2022-07-11 thomaspinder/GPJax#76 @thomaspinder
Hi @gpleiss.
I have resolved your issues now with a few follow up questions/comments.

- I have added a `[cuda]` flag to GPJax's requirements that installs `jax[cuda]`, as you suggested. Looking at other popular Jax-based libraries though, it seems as though most simply install regular Jax e.g., [jraph](https://github.com/deepmind/jraph/blob/master/setup.py) and [BrainPy](https://github.com/PKU-NIP-Lab/BrainPy/blob/master/setup.py), the latter having a cuda flag that isn't actioned. Despite this though, I agree with your suggestion of adding a cuda flag, however, in your experience with Torch/GPyTorch, has supporting such functionality ever caused headaches?
- Additional kernels and a high-dimensional example are currently being developed [here](https://github.com/thomaspinder/GPJax/tree/More_Kernels)
- My understanding was that the Github action in `.github/workflows/workflow-master.yml` runs unit tests on push and PR to master branch. Did you have something different in mind?
- I have now included a UCI regression benchmark task for the yacht dataset ([notebook here](https://gpjax.readthedocs.io/en/latest/nbs/yacht.html)).
- The paper has been refactored to better respond to the statement of need.
- I very much like your suggestion for larger end-to-end tests. I will sit down with Dan and sketch out what these may look like.

Thanks again for providing helpful comments as a reviewer.
