### 2022-07-13 #4455 @thomaspinder
Updating the pdf to reflect the updates made in response to the review from @gpleiss
