+++ gpjax/kernels.py
@@ -1,12 +1,13 @@
 import abc
-from logging import raiseExceptions
-from typing import Callable, Dict, List, Optional, Sequence
+from typing import Callable, Dict, List, Optional, Sequence, Tuple
 
+import distrax as dx
 import jax.numpy as jnp
 from chex import dataclass
 from jax import vmap
 
-from gpjax.types import Array
+from .config import Softplus, add_parameter
+from .types import Array
 
 
 ##########################################
@@ -171,12 +172,13 @@ class RBF(Kernel):
         K = params["variance"] * jnp.exp(-0.5 * squared_distance(x, y))
         return K.squeeze()
 
+
 @dataclass(repr=False)
 class PoweredExponential(Kernel):
     """The powered exponential family of kernels.
 
     Key reference is Diggle and Ribeiro (2007) - "Model-based Geostatistics".
-    
+
     """
 
     name: Optional[str] = "Powered exponential"
@@ -217,13 +219,13 @@ class Linear(Kernel):
 
     def __post_init__(self):
         self.ndims = 1 if not self.active_dims else len(self.active_dims)
-        self._params = {"variance": jnp.array([1.0]), "bias": jnp.array([0.0])}
+        self._params = {"variance": jnp.array([1.0])}
 
     def __call__(self, x: jnp.DeviceArray, y: jnp.DeviceArray, params: dict) -> Array:
-        """Evaluate the kernel on a pair of inputs :math:`(x, y)` with bias parameter :math:`b` and variance :math:`\sigma`
+        """Evaluate the kernel on a pair of inputs :math:`(x, y)` with variance parameter :math:`\sigma`
 
         .. math::
-            k(x, y) = \\sigma^2 x^{T}y + b
+            k(x, y) = \\sigma^2 x^{T}y
 
         Args:
             x (jnp.DeviceArray): The left hand argument of the kernel function's call.
@@ -234,7 +236,7 @@ class Linear(Kernel):
         """
         x = self.slice_input(x)
         y = self.slice_input(y)
-        K = params["variance"] * jnp.matmul(x.T, y) + params["bias"]
+        K = params["variance"] * jnp.matmul(x.T, y)
         return K.squeeze()
 
 
@@ -380,6 +382,7 @@ class Polynomial(Kernel):
         K = jnp.power(params["shift"] + jnp.dot(x * params["variance"], y), self.degree)
         return K.squeeze()
 
+
 @dataclass(repr=False)
 class RationalQuadratic(Kernel):
 
@@ -408,8 +411,11 @@ class RationalQuadratic(Kernel):
         """
         x = self.slice_input(x) / params["lengthscale"]
         y = self.slice_input(y) / params["lengthscale"]
-        K = params["variance"] * (1 +  0.5 * squared_distance(x, y) / params["alpha"]) ** (-params["alpha"])
+        K = params["variance"] * (
+            1 + 0.5 * squared_distance(x, y) / params["alpha"]
+        ) ** (-params["alpha"])
         return K.squeeze()
+
 
 @dataclass(repr=False)
 class Periodic(Kernel):
@@ -428,7 +434,6 @@ class Periodic(Kernel):
             "period": jnp.array([1.0] * self.ndims),
         }
 
-
     def __call__(self, x: jnp.DeviceArray, y: jnp.DeviceArray, params: dict) -> Array:
         """Evaluate the kernel on a pair of inputs :math:`(x, y)` with length-scale parameter :math:`\ell` and variance :math:`\sigma`
 
@@ -442,71 +447,97 @@ class Periodic(Kernel):
         Returns:
             Array: The value of :math:`k(x, y)`
         """
-        x = self.slice_input(x) 
+        x = self.slice_input(x)
         y = self.slice_input(y)
[…]
