### 2022-07-20 2a9e32d
Initial kronecker work
