### 2022-07-20 69bc705
Initial kronecker work
