+++ gpjax/kernels.py
@@ -147,6 +147,69 @@ class DenseKernelComputation(AbstractKernelComputation):
 
 
 @dataclass
+class AbstractKernelComputation:
+    @abc.abstractmethod
+    def gram():
+        raise NotImplementedError
+
+    @abc.abstractmethod
+    def diagonal():
+        raise NotImplementedError
+
+    @abc.abstractmethod
+    def cross_covariance():
+        raise NotImplementedError
+
+    def solve():
+        raise NotImplementedError
+
+
+class DenseKernelComputation(AbstractKernelComputation):
+    @staticmethod
+    def gram(kernel: Kernel, inputs: f64["N D"], params: dict) -> f64["N N"]:
+        """For a given kernel, compute the :math:`n \times n` gram matrix on an input matrix of shape :math:`n \times d` for :math:`d\geq 1`.
+
+        Args:
+            kernel (Kernel): The kernel for which the Gram matrix should be computed for.
+            inputs (Array): The input matrix.
+            params (dict): The kernel's parameter set.
+
+        Returns:
+            Array: The computed square Gram matrix.
+        """
+        return vmap(lambda x1: vmap(lambda y1: kernel(x1, y1, params))(inputs))(inputs)
+
+    @staticmethod
+    def cross_covariance(
+        kernel: Kernel, x: f64["N D"], y: f64["M D"], params: dict
+    ) -> f64["N M"]:
+        """For a given kernel, compute the :math:`m \times n` gram matrix on an a pair of input matrices with shape :math:`m \times d`  and :math:`n \times d` for :math:`d\geq 1`.
+
+        Args:
+            kernel (Kernel): The kernel for which the cross-covariance matrix should be computed for.
+            x (Array): The first input matrix.
+            y (Array): The second input matrix.
+            params (dict): The kernel's parameter set.
+
+        Returns:
+            Array: The computed square Gram matrix.
+        """
+        return vmap(lambda x1: vmap(lambda y1: kernel(x1, y1, params))(y))(x)
+
+    @staticmethod
+    def diagonal(kernel: Kernel, inputs: f64["N D"], params: dict) -> f64["N N"]:
+        """For a given kernel, compute the elementwise diagonal of the :math:`n \times n` gram matrix on an input matrix of shape :math:`n \times d` for :math:`d\geq 1`.
+        Args:
+            kernel (Kernel): The kernel for which the variance vector should be computed for.
+            inputs (Array): The input matrix.
+            params (dict): The kernel's parameter set.
+        Returns:
+            Array: The computed diagonal variance matrix.
+        """
+        return jnp.diag(vmap(lambda x: kernel(x, x, params))(inputs))
+
+
+@dataclass
 class _KernelSet:
     kernel_set: List[Kernel]
 
