### 2022-07-20 thomaspinder/GPJax#76 @gpleiss
Thanks for addressing my comments!

> I have added a [cuda] flag to GPJax's requirements that installs jax[cuda], as you suggested. Looking at other popular Jax-based libraries though, it seems as though most simply install regular Jax e.g., jraph and BrainPy, the latter having a cuda flag that isn't actioned. Despite this though, I agree with your suggestion of adding a cuda flag, however, in your experience with Torch/GPyTorch, has supporting such functionality ever caused headaches?

Hmm... I personally haven't noticed a headache with Torch/GPyTorch, and I don't think we've ever had any GitHub issues about cuda installation. There probably is a better pattern out there for cuda installation, but I don't think that this project has to figure it out 😄

My understanding was that the Github actio

> n in .github/workflows/workflow-master.yml runs unit tests on push and PR to master branch. Did you have something different in mind?

Oops! I must have missed this. The GitHub action is exactly what I had in mind.

> I have now included a UCI regression benchmark task for the yacht dataset (notebook here).

The notebook looks awesome! I really like the diagnostic plots (and I am definitely stealing the idea for our gpytorch tutorials 😉)

> I very much like your suggestion for larger end-to-end tests. I will sit down with Dan and sketch out what these may look like.

Awesome!
