+++ gpjax/variational_families.py
@@ -43,13 +43,16 @@ class AbstractVariationalFamily:
         """Predict the GP's output given the input."""
         raise NotImplementedError
 
+
 @dataclass
 class AbstractVariationalGaussian(AbstractVariationalFamily):
     """The variational Gaussian family of probability distributions."""
+
     prior: Prior
-    inducing_inputs: Array
+    inducing_inputs: f64["N D"]
     name: str = "Gaussian"
     jitter: Optional[float] = DEFAULT_JITTER
+
 
 @dataclass
 class VariationalGaussian(AbstractVariationalGaussian):
@@ -59,10 +62,10 @@ class VariationalGaussian(AbstractVariationalGaussian):
     and the distribution over the inducing inputs is q(u) = N(μ, S). We parameterise this over μ and sqrt with S = sqrt sqrtᵀ.
 
     """
-    variational_mean: Optional[Array] = None
-    variational_root_covariance: Optional[Array] = None
+
+    variational_mean: Optional[f64["N D"]] = None
+    variational_root_covariance: Optional[f64["N D"]] = None
     diag: Optional[bool] = False
-    
 
     def __post_init__(self):
         """Initialise the variational Gaussian distribution."""
@@ -85,13 +88,16 @@ class VariationalGaussian(AbstractVariationalGaussian):
     def _initialise_params(self, key: jnp.DeviceArray) -> Dict:
         """Return the variational mean vector, variational root covariance matrix, and inducing input vector that parameterise the variational Gaussian distribution."""
         return concat_dictionaries(
-            self.prior._initialise_params(key), {
-            "variational_family": {
-                "inducing_inputs": self.inducing_inputs,
-                    "moments": {"variational_mean": self.variational_mean,
-                    "variational_root_covariance": self.variational_root_covariance}
+            self.prior._initialise_params(key),
+            {
+                "variational_family": {
+                    "inducing_inputs": self.inducing_inputs,
+                    "moments": {
+                        "variational_mean": self.variational_mean,
+                        "variational_root_covariance": self.variational_root_covariance,
+                    },
                 }
-            }
+            },
         )
 
     def prior_kl(self, params: Dict) -> Float[Array, "1"]:
@@ -104,7 +110,7 @@ class VariationalGaussian(AbstractVariationalGaussian):
             params (Dict): The parameters at which our variational distribution and GP prior are to be evaluated.
 
         Returns:
-            Array: The KL-divergence between our variational approximation and the GP prior.
+            f64["1"]: The KL-divergence between our variational approximation and the GP prior.
         """
         mu = params["variational_family"]["moments"]["variational_mean"]
         sqrt = params["variational_family"]["moments"]["variational_root_covariance"]
@@ -131,7 +137,7 @@ class VariationalGaussian(AbstractVariationalGaussian):
             params (dict): The set of parameters that are to be used to parameterise our variational approximation and GP.
 
         Returns:
-            Callable[[Array], dx.Distribution]: A function that accepts a set of test points and will return the predictive distribution at those points.
+            Callable[[f64["N D"]], dx.Distribution]: A function that accepts a set of test points and will return the predictive distribution at those points.
         """
         mu = params["variational_family"]["moments"]["variational_mean"]
         sqrt = params["variational_family"]["moments"]["variational_root_covariance"]
@@ -197,7 +203,7 @@ class WhitenedVariationalGaussian(VariationalGaussian):
[…]
