### 2022-07-22 #4455 @dfm
@thomaspinder — I've opened a PR with some minor edits to the paper. After merging or responding to that, here are the final steps that I'll need from you:
