### 2022-07-24 7b64000
Add logo to docs
