+++ docs/index.rst
@@ -6,7 +6,7 @@ Welcome to GPJax!
 
 GPJax is a didactic Gaussian process library that supports GPU acceleration and just-in-time compilation. We seek to provide a flexible API as close as possible to how the underlying mathematics is written on paper to enable researchers to rapidly prototype and develop new ideas.
 
-.. image:: ./GP.pdf
+.. image:: ./gpjax_log.pdf
   :width: 800
   :alt: Alternative text
 
