### 2022-07-25 bb618e2
Add docstrings to all props
