### 2022-07-25 7afb1bd
Replace text with logo
