### 2022-07-25 #4455 @kyleniemeyer
Hi @thomaspinder, I'm doing some final checks before publishing. It looks like a few references are missing DOIs (not sure why `editorialbot` didn't catch them), specifically for articles from the _Journal of Statistical Software_. Could you add those? I would also point out that arXiv articles also now have DOIs that can/probably should be included, though we are not currently able to automatically check for those.
