+++ paper.bib
@@ -3,7 +3,8 @@
   author    = {Rasmussen, Carl Edward and Williams, Christopher K},
   number    = {3},
   year      = {2006},
-  publisher = {MIT press Cambridge, MA}
+  publisher = {MIT press Cambridge, MA},
+  doi = {https://doi.org/10.7551/mitpress/3206.001.0001}
 }
 
 @article{matthews2017gpflow,
@@ -59,11 +60,12 @@
 @article{carpenter2017stan,
   title     = {Stan: A probabilistic programming language},
   author    = {Carpenter, Bob and Gelman, Andrew and Hoffman, Matthew D and Lee, Daniel and Goodrich, Ben and Betancourt, Michael and Brubaker, Marcus and Guo, Jiqiang and Li, Peter and Riddell, Allen},
-  journal   = {Journal of statistical software},
+  journal   = {Journal of Statistical Software},
   volume    = {76},
   number    = {1},
   year      = {2017},
-  publisher = {Columbia Univ., New York, NY (United States); Harvard Univ., Cambridge, MA~…}
+  publisher = {Columbia Univ., New York, NY (United States); Harvard Univ., Cambridge, MA~…},
+  doi = {10.18637/jss.v076.i01}
 }
 
 
@@ -88,12 +90,16 @@
 }
 
 @article{fairbrother2022gaussianprocesses,
-  title   = {{GaussianProcesses.jl}: A Nonparametric {Bayes} Package for the {Julia} Language},
-  author  = {Fairbrother, Jamie and Nemeth, Christopher and Rischard, Maxime and Brea, Johanni and Pinder, Thomas},
-  journal = {Journal of Statistical Software},
-  volume  = {102},
-  pages   = {1--36},
-  year    = {2022}
+  title={GaussianProcesses.jl: A Nonparametric Bayes Package for the Julia Language},
+  volume={102},
+  url={https://www.jstatsoft.org/index.php/jss/article/view/v102i01},
+  doi={10.18637/jss.v102.i01},
+  abstract={Gaussian processes are a class of flexible nonparametric Bayesian tools that are widely used across the sciences, and in industry, to model complex data sources. Key to applying Gaussian process models is the availability of well-developed open source software, which is available in many programming languages. In this paper, we present a tutorial of the GaussianProcesses.jl package that has been developed for the Julia programming language. GaussianProcesses.jl utilizes the inherent computational benefits of the Julia language, including multiple dispatch and just-in-time compilation, to produce a fast, flexible and user-friendly Gaussian processes package. The package provides many mean and kernel functions with supporting inference tools to fit exact Gaussian process models, as well as a range of alternative likelihood functions to handle non-Gaussian data (e.g., binary classification models) and sparse approximations for scalable Gaussian processes. The package makes efficient use of existing Julia packages to provide users with a range of optimization and plotting tools.},
+  number={1},
+  journal={Journal of Statistical Software},
+  author={Fairbrother, Jamie and Nemeth, Christopher and Rischard, Maxime and Brea, Johanni and Pinder, Thomas},
+  year={2022},
+  pages={1–36}
 }
 
 @article{vanhatalo2013gpstuff,
@@ -171,11 +177,18 @@
 }
 
 @article{hensman2013gaussian,
-  title   = {{G}aussian processes for big data},
-  author  = {Hensman, James and Fusi, Nicolo and Lawrence, Neil D},
-  journal = {arXiv preprint arXiv:1309.6835},
-  year    = {2013}
+  author    = {James Hensman and
+               Nicol{\'{o}} Fusi and
[…]
