+++ paper.bib
@@ -4,7 +4,7 @@
   number    = {3},
   year      = {2006},
   publisher = {MIT press Cambridge, MA},
-  doi = {https://doi.org/10.7551/mitpress/3206.001.0001}
+  doi = {10.7551/mitpress/3206.001.0001}
 }
 
 @article{matthews2017gpflow,
