+++ paper.bib
@@ -54,7 +54,8 @@
   author  = {Phan, Du and Pradhan, Neeraj and Jankowiak, Martin},
   title   = {Composable Effects for Flexible and Accelerated Probabilistic Programming in {NumPyro}},
   journal = {arXiv preprint arXiv:1912.11554},
-  year    = {2019}
+  year    = {2019},
+  doi = {10.48550/arxiv.1912.11554}
 }
 
 @article{carpenter2017stan,
@@ -86,6 +87,7 @@
   title   = {{B}ayes-{N}ewton Methods for Approximate {B}ayesian Inference with {PSD} Guarantees},
   author  = {Wilkinson, William J. and S\"arkk\"a, Simo and Solin, Arno},
   journal = {arXiv preprint arXiv:2111.01721},
+  doi = {10.48550/arxiv.2111.01721},
   year    = {2021}
 }
 
@@ -120,6 +122,7 @@
   title   = {{Gaussian} Processes on Hypergraphs},
   author  = {Pinder, Thomas and Turnbull, Kathryn and Nemeth, Christopher and Leslie, David},
   journal = {arXiv preprint arXiv:2106.01982},
+  doi = {10.48550/arxiv.2106.01982},
   year    = {2021}
 }
 
