### 2022-07-26 #4455 @thomaspinder
Thanks for catching that @kyleniemeyer. All the DOIs should be accounted for now, including arXiv references.
